@@ -3,9 +3,15 @@ package org.soak.wrapper;
 import com.destroystokyo.paper.MaterialSetTag;
 import com.destroystokyo.paper.entity.ai.MobGoals;
 import com.destroystokyo.paper.profile.PlayerProfile;
+import io.papermc.paper.ban.BanListType;
 import io.papermc.paper.datapack.DatapackManager;
+import io.papermc.paper.math.Position;
 import io.papermc.paper.tag.EntitySetTag;
+import io.papermc.paper.threadedregions.scheduler.AsyncScheduler;
+import io.papermc.paper.threadedregions.scheduler.GlobalRegionScheduler;
+import io.papermc.paper.threadedregions.scheduler.RegionScheduler;
 import net.kyori.adventure.audience.Audience;
+import net.kyori.adventure.key.Key;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
@@ -17,17 +23,16 @@ import org.bukkit.command.CommandException;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.ConsoleCommandSender;
 import org.bukkit.command.PluginCommand;
-import org.bukkit.entity.Entity;
-import org.bukkit.entity.EntityType;
-import org.bukkit.entity.Player;
-import org.bukkit.entity.SpawnCategory;
+import org.bukkit.entity.*;
 import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.generator.structure.Structure;
 import org.bukkit.inventory.*;
 import org.bukkit.loot.LootTable;
+import org.bukkit.map.MapCursor;
 import org.bukkit.map.MapView;
 import org.bukkit.packs.DataPackManager;
+import org.bukkit.packs.ResourcePack;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.ServicesManager;
 import org.bukkit.plugin.SimplePluginManager;
@@ -44,9 +49,8 @@ import org.jetbrains.annotations.Nullable;
 import org.mose.collection.stream.builder.CollectionStreamBuilder;
 import org.soak.WrapperManager;
 import org.soak.exception.NotImplementedException;
-import org.soak.map.SoakGameModeMap;
-import org.soak.map.SoakMessageMap;
-import org.soak.map.SoakResourceKeyMap;
+import org.soak.map.*;
+import org.soak.map.item.SoakItemStackMap;
 import org.soak.map.item.SoakRecipeMap;
 import org.soak.plugin.SoakExternalManager;
 import org.soak.plugin.SoakManager;
@@ -74,6 +78,7 @@ import org.spongepowered.api.entity.living.player.User;
 import org.spongepowered.api.item.ItemType;
 import org.spongepowered.api.item.ItemTypes;
 import org.spongepowered.api.item.inventory.type.ViewableInventory;
+import org.spongepowered.api.profile.GameProfile;
 import org.spongepowered.api.registry.RegistryTypes;
 import org.spongepowered.api.service.ban.Ban;
 import org.spongepowered.api.service.ban.BanTypes;
@@ -101,7 +106,12 @@ import java.util.logging.ConsoleHandler;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
 
-public abstract class SoakServer implements SimpServer {
+/*
+    This class is a bloated mess.
+    Maybe worth splitting this out into their own managers and then using this
+    class to redirect the call to its respected manager.
+ */
+public abstract class SoakServer implements Server {
 
     private final Supplier<org.spongepowered.api.Server> serverSupplier;
     private final Singleton<SoakMessenger> messenger = new Singleton<>(() -> new SoakMessenger(Sponge.channelManager()));
@@ -119,30 +129,21 @@ public abstract class SoakServer implements SimpServer {
         pm.paperPluginManager = pluginManager.get();
         return pm;
     });
-    private final Singleton<Map<Class<?>, Registry<?>>> registries = new Singleton<>(() -> Arrays
-            .stream(Registry.class.getDeclaredFields())
-            .filter(field -> Modifier.isFinal(field.getModifiers()))
-            .filter(field -> Modifier.isPublic(field.getModifiers()))
-            .filter(field -> Modifier.isStatic(field.getModifiers()))
-            .filter(field -> Registry.class.isAssignableFrom(field.getType()))
-            .map(field -> {
-                try {
-                    var reg = (Registry<?>) field.get(null);
-                    return reg;
-                } catch (IllegalAccessException e) {
-                    throw new RuntimeException(e);
-                }
-            })
-            .filter(Objects::nonNull)
-            .filter(reg -> !(reg instanceof SoakRegistry<?, ?>))
-            .collect(Collectors.toMap(reg -> {
-                if (reg instanceof Registry.SimpleRegistry simpleReg) {
-                    var first = simpleReg.iterator().next();
-                    return first.getClass();
-                }
-                var generic = (ParameterizedType) reg.getClass().getGenericInterfaces()[0];
-                return (Class<?>) generic.getActualTypeArguments()[0];
-            }, reg -> reg)));
+    private final Singleton<Map<Class<?>, Registry<?>>> registries = new Singleton<>(() -> Arrays.stream(Registry.class.getDeclaredFields()).filter(field -> Modifier.isFinal(field.getModifiers())).filter(field -> Modifier.isPublic(field.getModifiers())).filter(field -> Modifier.isStatic(field.getModifiers())).filter(field -> Registry.class.isAssignableFrom(field.getType())).map(field -> {
+        try {
+            var reg = (Registry<?>) field.get(null);
+            return reg;
+        } catch (IllegalAccessException e) {
+            throw new RuntimeException(e);
+        }
+    }).filter(Objects::nonNull).filter(reg -> !(reg instanceof SoakRegistry<?, ?>)).collect(Collectors.toMap(reg -> {
+        if (reg instanceof Registry.SimpleRegistry simpleReg) {
+            var first = simpleReg.iterator().next();
+            return first.getClass();
+        }
+        var generic = (ParameterizedType) reg.getClass().getGenericInterfaces()[0];
+        return (Class<?>) generic.getActualTypeArguments()[0];
+    }, reg -> reg)));
 
     private final java.util.logging.Logger logger;
 
@@ -156,60 +157,222 @@ public abstract class SoakServer implements SimpServer {
         this.serverSupplier = serverSupplier;
     }
 
+    @Override
+    public boolean isAcceptingTransfers() {
+        throw NotImplementedException.createByLazy(Server.class, "isAcceptingTransfers");
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull World world, @NotNull Position position) {
+        return isOwnedByCurrentRegion(position.toLocation(world));
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull World world, @NotNull Position position, int squareRadiusChecks) {
+        return isOwnedByCurrentRegion(position.toLocation(world), squareRadiusChecks);
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull Location location) {
+        return isOwnedByCurrentRegion(location, 1);
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull Location location, int squareRadiusChunks) {
+        var chunk = location.getChunk();
+        return isOwnedByCurrentRegion(location.getWorld(), chunk.getX(), chunk.getZ(), squareRadiusChunks);
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull World world, int chunkX, int chunkZ) {
+        return isOwnedByCurrentRegion(world, chunkX, chunkZ, 1);
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull World world, int chunkX, int chunkZ, int squareRadiusChunks) {
+        throw NotImplementedException.createByLazy(Server.class, "isOwnedByCurrentRegion", World.class, int.class, int.class, int.class);
+    }
+
+    @Override
+    public boolean isOwnedByCurrentRegion(@NotNull Entity entity) {
+        return isOwnedByCurrentRegion(entity.getLocation());
+    }
+
+    @Override
+    public @NotNull RegionScheduler getRegionScheduler() {
+        throw NotImplementedException.createByLazy(Server.class, "getRegionScheduler");
+    }
+
+    @Override
+    public @NotNull AsyncScheduler getAsyncScheduler() {
+        throw NotImplementedException.createByLazy(Server.class, "getAsyncScheduler");
+    }
+
+    @Override
+    public @NotNull GlobalRegionScheduler getGlobalRegionScheduler() {
+        throw NotImplementedException.createByLazy(Server.class, "getGlobalRegionScheduler");
+    }
+
+    @Override
+    public void banIP(@NotNull InetAddress inetAddress) {
+        throw NotImplementedException.createByLazy(Server.class, "banIP", InetAddress.class);
+    }
+
+    @Override
+    public void unbanIP(@NotNull InetAddress inetAddress) {
+        throw NotImplementedException.createByLazy(Server.class, "unbanIP", InetAddress.class);
+    }
+
+    @Override
+    public <B extends BanList<E>, E> @NotNull B getBanList(@NotNull BanListType<B> banListType) {
+        throw NotImplementedException.createByLazy(Server.class, "getBanList", BanListType.class);
+    }
+
+    @Override
+    public @NotNull EntityFactory getEntityFactory() {
+        throw NotImplementedException.createByLazy(Server.class, "getEntityFactory");
+    }
+
+    @Override
+    public @NotNull ServerLinks getServerLinks() {
+        throw NotImplementedException.createByLazy(Server.class, "getServerLinks");
+    }
+
+    @Override
+    public @NotNull ItemStack craftItem(@NotNull ItemStack[] itemStacks, @NotNull World world) {
+        throw NotImplementedException.createByLazy(Server.class, "craftItem", ItemStack[].class, World.class);
+    }
+
+    @Override
+    public @NotNull ItemCraftResult craftItemResult(@NotNull ItemStack[] itemStacks, @NotNull World world, @NotNull Player player) {
+        throw NotImplementedException.createByLazy(Server.class, "craftItemResult", ItemStack[].class, World.class, Player.class);
+    }
+
+    @Override
+    public @NotNull ItemCraftResult craftItemResult(@NotNull ItemStack[] itemStacks, @NotNull World world) {
+        throw NotImplementedException.createByLazy(Server.class, "craftItemResult", ItemStack[].class, World.class);
+    }
+
+    @Override
+    public void updateResources() {
+        throw NotImplementedException.createByLazy(Server.class, "updateResources");
+    }
+
+    @Override
+    public void updateRecipes() {
+        throw NotImplementedException.createByLazy(Server.class, "updateRecipes");
+    }
+
+    @Override
+    public boolean removeRecipe(@NotNull NamespacedKey namespacedKey, boolean b) {
+        throw NotImplementedException.createByLazy(Server.class, "removeRecipe", NamespacedKey.class, boolean.class);
+    }
+
+    @Override
+    public @Nullable ItemStack createExplorerMap(@NotNull World world, @NotNull Location location, @NotNull org.bukkit.generator.structure.StructureType structureType, @NotNull MapCursor.Type type, int i, boolean b) {
+        Bukkit bukkit;
+        throw NotImplementedException.createByLazy(Server.class, "createExplorerMap", World.class, Location.class, org.bukkit.generator.structure.StructureType.class, MapCursor.Type.class, int.class, boolean.class);
+    }
+
     public org.spongepowered.api.Server spongeServer() {
         return this.serverSupplier.get();
     }
 
     @Override
+    public @Nullable Recipe getRecipe(@NotNull NamespacedKey namespacedKey) {
+        throw NotImplementedException.createByLazy(Server.class, "getRecipe", NamespacedKey.class);
+    }
+
+    @Override
+    public @NotNull List<Recipe> getRecipesFor(@NotNull ItemStack itemStack) {
+        throw NotImplementedException.createByLazy(Server.class, "getRecipesFor", ItemStack.class);
+    }
+
+    @Override
+    public boolean unloadWorld(@NotNull String name, boolean save) {
+        var world = getWorld(name);
+        if (world == null) {
+            return false;
+        }
+        return unloadWorld(world, save);
+    }
+
+    @Override
+    public @Nullable World getWorld(@NotNull Key key) {
+        var namespace = NamespacedKey.fromString(key.asString());
+        if (namespace == null) {
+            return null;
+        }
+        return this.getWorld(namespace);
+    }
+
+    @Override
+    public @Nullable World getWorld(@NotNull String s) {
+        return Sponge.server().worldManager().worlds().stream().filter(world -> world.key().value().equalsIgnoreCase(s)).findAny().map(world -> SoakManager.<WrapperManager>getManager().getMemoryStore().get(world)).orElse(null);
+    }
+
+    @Override
+    public @Nullable World getWorld(@NotNull UUID uuid) {
+        var worldManager = Sponge.server().worldManager();
+        return worldManager.worldKey(uuid).flatMap(worldManager::world).map(world -> SoakManager.<WrapperManager>getManager().getMemoryStore().get(world)).orElse(null);
+    }
+
+    @Override
+    public @Nullable ResourcePack getServerResourcePack() {
+        throw NotImplementedException.createByLazy(Server.class, "getServerResourcePack");
+    }
+
+    @Override
+    public boolean isLoggingIPs() {
+        throw NotImplementedException.createByLazy(Server.class, "isLoggingIPs");
+    }
+
+    @Override
+    public @NotNull ServerTickManager getServerTickManager() {
+        throw NotImplementedException.createByLazy(Server.class, "getServerTickManager");
+    }
+
+    @Override
+    public @NotNull <T extends Keyed> Iterable<Tag<T>> getTags(@NotNull String registry, @NotNull Class<T> clazz) {
+        throw NotImplementedException.createByLazy(Server.class, "getTags", String.class, Class.class);
+    }
+
+    @Override
     public <T extends Keyed> Tag<T> getTag(@NotNull String registry, @NotNull NamespacedKey tag, @NotNull Class<T> clazz) {
         ResourceKey key = SoakResourceKeyMap.mapToSponge(tag);
-        if (clazz.getName().equals(Material.class.getName()) && registry.equals(Tag.REGISTRY_BLOCKS)) {
-            var opTag = FakeRegistryHelper.<org.spongepowered.api.tag.Tag<BlockType>>getFields(BlockTypeTags.class, org.spongepowered.api.tag.Tag.class)
-                    .stream()
-                    .filter(spongeTag -> spongeTag.key().equals(key))
-                    .findAny();
+        boolean classMatch = clazz.getName().equals(Material.class.getName());
+        if (classMatch && registry.equals(Tag.REGISTRY_BLOCKS)) {
+            var opTag = FakeRegistryHelper.<org.spongepowered.api.tag.Tag<BlockType>>getFields(BlockTypeTags.class, org.spongepowered.api.tag.Tag.class).stream().filter(spongeTag -> spongeTag.key().equals(key)).findAny();
             if (opTag.isPresent()) {
-                return (Tag<T>) (Object) new MaterialSetTag(tag, TagHelper.getBlockTypes(opTag.get()).map(Material::getBlockMaterial).collect(Collectors.toList()));
+                return (Tag<T>) (Object) new MaterialSetTag(tag, TagHelper.getBlockTypes(opTag.get()).map(block -> SoakBlockMap.toBukkit(block)).collect(Collectors.toList()));
             }
         }
-        if (clazz.getName().equals(Material.class.getName()) && registry.equals(Tag.REGISTRY_ITEMS)) {
-            var opTag = FakeRegistryHelper.<org.spongepowered.api.tag.Tag<ItemType>>getFields(ItemTypeTags.class, org.spongepowered.api.tag.Tag.class)
-                    .stream()
-                    .filter(spongeTag -> spongeTag.key().equals(key))
-                    .findAny();
+        if (classMatch && registry.equals(Tag.REGISTRY_ITEMS)) {
+            var opTag = FakeRegistryHelper.<org.spongepowered.api.tag.Tag<ItemType>>getFields(ItemTypeTags.class, org.spongepowered.api.tag.Tag.class).stream().filter(spongeTag -> spongeTag.key().equals(key)).findAny();
             if (opTag.isPresent()) {
-                return (Tag<T>) (Object) new MaterialSetTag(tag, TagHelper.getItemTypes(opTag.get()).map(Material::getItemMaterial).collect(Collectors.toList()));
+                return (Tag<T>) (Object) new MaterialSetTag(tag, TagHelper.getItemTypes(opTag.get()).map(item -> SoakItemStackMap.toBukkit(item)).collect(Collectors.toList()));
             }
         }
         if (clazz.getName().equals(EntityType.class.getName()) && registry.equals(Tag.REGISTRY_ENTITY_TYPES)) {
-            var opTag = FakeRegistryHelper.<org.spongepowered.api.tag.Tag<org.spongepowered.api.entity.EntityType<?>>>getFields(EntityTypeTags.class, org.spongepowered.api.tag.Tag.class)
-                    .stream()
-                    .filter(spongeTag -> spongeTag.key().equals(key))
-                    .findAny();
+            var opTag = FakeRegistryHelper.<org.spongepowered.api.tag.Tag<org.spongepowered.api.entity.EntityType<?>>>getFields(EntityTypeTags.class, org.spongepowered.api.tag.Tag.class).stream().filter(spongeTag -> spongeTag.key().equals(key)).findAny();
             if (opTag.isPresent()) {
-                return (Tag<T>) (Object) new EntitySetTag(tag, TagHelper.getEntityTypes(opTag.get()).map(EntityType::fromSponge).collect(Collectors.toList()));
+                return (Tag<T>) (Object) new EntitySetTag(tag, TagHelper.getEntityTypes(opTag.get()).map(entity -> SoakEntityMap.toBukkit(entity)).collect(Collectors.toList()));
             }
         }
 
 
         //overrides
         if (registry.equals(Tag.REGISTRY_ITEMS) && tag.asString().equals("minecraft:coral_blocks")) {
-            return (Tag<T>) (Object) new MaterialSetTag(tag,
-                    ItemTypes.registry()
-                            .stream()
-                            .filter(item -> item.key(RegistryTypes.ITEM_TYPE).value().contains("coral_blocks"))
-                            .map(item -> Material.getItemMaterial(item))
-                            .collect(Collectors.toList()));
+            return (Tag<T>) (Object) new MaterialSetTag(tag, ItemTypes.registry().stream().filter(item -> item.key(RegistryTypes.ITEM_TYPE).value().contains("coral_blocks")).map(item -> SoakItemStackMap.toBukkit(item)).collect(Collectors.toList()));
         }
 
-        if (registry.equals(Tag.REGISTRY_BLOCKS) && tag.asString()
-                .equals("minecraft:wool_carpets")) {
-            Set<Material> itemTypes = TagHelper.getBlockTypes(BlockTypeTags.WOOL_CARPETS).map(Material::getBlockMaterial).collect(Collectors.toSet());
+        if (registry.equals(Tag.REGISTRY_BLOCKS) && tag.asString().equals("minecraft:wool_carpets")) {
+            Set<Material> itemTypes = TagHelper.getBlockTypes(BlockTypeTags.WOOL_CARPETS).map(block -> SoakBlockMap.toBukkit(block)).collect(Collectors.toSet());
             return (Tag<T>) (Object) new MaterialSetTag(tag, itemTypes);
         }
 
         if (registry.equals(Tag.REGISTRY_ITEMS) && tag.asString().equals("minecraft:wool_carpets")) {
-            Set<Material> itemTypes = TagHelper.getItemTypes(ItemTypeTags.WOOL_CARPETS).map(Material::getItemMaterial).collect(Collectors.toSet());
+            Set<Material> itemTypes = TagHelper.getItemTypes(ItemTypeTags.WOOL_CARPETS).map(item -> SoakItemStackMap.toBukkit(item)).collect(Collectors.toSet());
             return (Tag<T>) (Object) new MaterialSetTag(tag, itemTypes);
 
         }
@@ -223,23 +386,11 @@ public abstract class SoakServer implements SimpServer {
         }
 
         if (registry.equals(Tag.REGISTRY_ITEMS) && tag.asString().equals("minecraft:crops")) {
-            var items = ItemTypes.registry()
-                    .stream()
-                    .filter(item -> org.spongepowered.api.item.inventory.ItemStack.of(item)
-                            .get(Keys.REPLENISHED_FOOD)
-                            .isPresent())
-                    .map(Material::getItemMaterial)
-                    .collect(Collectors.toSet());
+            var items = ItemTypes.registry().stream().filter(item -> org.spongepowered.api.item.inventory.ItemStack.of(item).get(Keys.REPLENISHED_FOOD).isPresent()).map(item -> SoakItemStackMap.toBukkit(item)).collect(Collectors.toSet());
             return (Tag<T>) (Object) new MaterialSetTag(tag, items);
         }
         if (registry.equals(Tag.REGISTRY_ITEMS) && tag.asString().equals("minecraft:furnace_materials")) {
-            var items = ItemTypes.registry()
-                    .stream()
-                    .filter(item -> org.spongepowered.api.item.inventory.ItemStack.of(item)
-                            .get(Keys.MAX_COOK_TIME)
-                            .isPresent())
-                    .map(Material::getItemMaterial)
-                    .collect(Collectors.toSet());
+            var items = ItemTypes.registry().stream().filter(item -> org.spongepowered.api.item.inventory.ItemStack.of(item).get(Keys.MAX_COOK_TIME).isPresent()).map(item -> SoakItemStackMap.toBukkit(item)).collect(Collectors.toSet());
             return (Tag<T>) (Object) new MaterialSetTag(tag, items);
         }
         SoakManager.getManager().getLogger().warn("No tag found of registry: '" + registry + "' Type: " + clazz.getSimpleName() + " id: " + tag.asString());
@@ -253,7 +404,7 @@ public abstract class SoakServer implements SimpServer {
     }
 
     @Override
-    public @NotNull Spigot spigot() {
+    public @NotNull Server.Spigot spigot() {
         throw NotImplementedException.createByLazy(Server.class, "spigot");
     }
 
@@ -264,7 +415,7 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @NotNull File getPluginsFolder() {
-        return new File("bukkit/plugins");
+        return new File("org/soak/generate/bukkit/plugins");
     }
 
     @Override
@@ -274,8 +425,7 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @NotNull String getVersion() {
-        return "(MC: " + getMinecraftVersion() + ") - " + SoakManager.getManager().getVersion()
-                .toString();
+        return "(MC: " + getMinecraftVersion() + ") - " + SoakManager.getManager().getVersion().toString();
     }
 
     @Override
@@ -293,17 +443,8 @@ public abstract class SoakServer implements SimpServer {
         if (!Sponge.isServerAvailable()) {
             return Collections.emptySet();
         }
-        var builder = CollectionStreamBuilder
-                .builder()
-                .collection(Sponge.server().onlinePlayers())
-                .basicMap(player -> (Player) SoakManager.<WrapperManager>getManager().getMemoryStore().get(player));
-        return ListMappingUtils
-                .fromStream(
-                        builder,
-                        () -> Sponge.server().onlinePlayers().stream(),
-                        (spongePlayer, soakPlayer) -> ((SoakPlayer) soakPlayer).spongeEntity().equals(spongePlayer),
-                        Comparator.comparing(spongePlayer -> spongePlayer.get(Keys.LAST_DATE_JOINED).orElseThrow(() -> new RuntimeException("No value found for last joined on a online player"))))
-                .buildList();
+        var builder = CollectionStreamBuilder.builder().collection(Sponge.server().onlinePlayers()).basicMap(player -> (Player) SoakManager.<WrapperManager>getManager().getMemoryStore().get(player));
+        return ListMappingUtils.fromStream(builder, () -> Sponge.server().onlinePlayers().stream(), (spongePlayer, soakPlayer) -> ((SoakPlayer) soakPlayer).spongeEntity().equals(spongePlayer), Comparator.comparing(spongePlayer -> spongePlayer.get(Keys.LAST_DATE_JOINED).orElseThrow(() -> new RuntimeException("No value found for last joined on a online player")))).buildList();
     }
 
     @Override
@@ -417,20 +558,13 @@ public abstract class SoakServer implements SimpServer {
         var userManager = this.spongeServer().userManager();
         var futurePlayers = whitelistService.whitelistedProfiles().thenCompose(profiles -> {
             var futures = profiles.stream().map(userManager::load).toArray(CompletableFuture[]::new);
-            return CompletableFuture
-                    .allOf(futures)
-                    .thenApply(v -> Arrays
-                            .stream(futures)
-                            .map(future -> {
-                                try {
-                                    return (Optional<User>) future.get();
-                                } catch (InterruptedException | ExecutionException e) {
-                                    throw new RuntimeException(e);
-                                }
-                            })
-                            .filter(Optional::isPresent)
-                            .map(Optional::get)
-                            .map(user -> (OfflinePlayer) new SoakOfflinePlayer(user)).collect(Collectors.toSet()));
+            return CompletableFuture.allOf(futures).thenApply(v -> Arrays.stream(futures).map(future -> {
+                try {
+                    return (Optional<User>) future.get();
+                } catch (InterruptedException | ExecutionException e) {
+                    throw new RuntimeException(e);
+                }
+            }).filter(Optional::isPresent).map(Optional::get).map(user -> (OfflinePlayer) new SoakOfflinePlayer(user)).collect(Collectors.toSet()));
         });
         try {
             return futurePlayers.get();
@@ -520,6 +654,17 @@ public abstract class SoakServer implements SimpServer {
         return player.getUniqueId();
     }
 
+    @Override
+    public @Nullable Player getPlayer(@NotNull UUID uuid) {
+        return Sponge.server().player(uuid).map(player -> SoakManager.<WrapperManager>getManager().getMemoryStore().get(player)).orElse(null);
+    }
+
+    @Override
+    public @Nullable Player getPlayer(@NotNull String s) {
+        return Sponge.server().player(s).map(player -> SoakManager.<WrapperManager>getManager().getMemoryStore().get(player)).orElse(null);
+
+    }
+
     public @NotNull SoakPluginManager getSoakPluginManager() {
         return this.pluginManager.get();
     }
@@ -541,12 +686,8 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @NotNull List<World> getWorlds() {
-        var builder = CollectionStreamBuilder
-                .builder()
-                .collection(this.spongeServer().worldManager().worlds())
-                .basicMap(world -> (World) SoakManager.<WrapperManager>getManager().getMemoryStore().get(world));
-        return ListMappingUtils.fromStream(builder, () -> this.spongeServer().worldManager().worlds().stream(), (spongeWorld, soakWorld) -> ((SoakWorld) soakWorld).sponge().equals(spongeWorld), Comparator.comparing(world -> world.key().formatted()))
-                .buildList();
+        var builder = CollectionStreamBuilder.builder().collection(this.spongeServer().worldManager().worlds()).basicMap(world -> (World) SoakManager.<WrapperManager>getManager().getMemoryStore().get(world));
+        return ListMappingUtils.fromStream(builder, () -> this.spongeServer().worldManager().worlds().stream(), (spongeWorld, soakWorld) -> ((SoakWorld) soakWorld).sponge().equals(spongeWorld), Comparator.comparing(world -> world.key().formatted())).buildList();
     }
 
     @Override
@@ -594,27 +735,17 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @NotNull ItemStack createExplorerMap(@NotNull World world, @NotNull Location location, @NotNull StructureType structureType) {
-        throw NotImplementedException.createByLazy(SoakServer.class,
-                "createExplorerMap",
-                World.class,
-                Location.class,
-                StructureType.class);
+        throw NotImplementedException.createByLazy(SoakServer.class, "createExplorerMap", World.class, Location.class, StructureType.class);
     }
 
     @Override
     public @NotNull ItemStack createExplorerMap(@NotNull World world, @NotNull Location location, @NotNull StructureType structureType, int radius, boolean findUnexplored) {
-        throw NotImplementedException.createByLazy(SoakServer.class,
-                "createExplorerMap",
-                World.class,
-                Location.class,
-                StructureType.class,
-                boolean.class);
+        throw NotImplementedException.createByLazy(SoakServer.class, "createExplorerMap", World.class, Location.class, StructureType.class, boolean.class);
     }
 
     @Override
     public void reload() {
-        SoakManager.getManager().getLogger()
-                .warn("A Bukkit plugin attempted to reload the plugin list. This is not possible in Sponge, reloading the data instead");
+        SoakManager.getManager().getLogger().warn("A Bukkit plugin attempted to reload the plugin list. This is not possible in Sponge, reloading the data instead");
         reloadData();
     }
 
@@ -625,19 +756,12 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @Nullable PluginCommand getPluginCommand(@NotNull String name) {
-        PluginCommand pluginCommand = SoakManager.getManager().getBukkitContainers()
-                .sorted(Comparator.comparing(pl -> pl.metadata().id()))
-                .flatMap(pl -> SoakManager.<WrapperManager>getManager().getBukkitCommands(pl.getBukkitInstance()).stream())
-                .filter(cmd -> {
-                    if (cmd.getName().equalsIgnoreCase(name)) {
-                        return true;
-                    }
-                    return cmd.getAliases().stream().anyMatch(alias -> alias.equalsIgnoreCase(name));
-                })
-                .filter(cmd -> cmd instanceof PluginCommand)
-                .findFirst()
-                .map(cmd -> (PluginCommand) cmd)
-                .orElse(null);
+        PluginCommand pluginCommand = SoakManager.getManager().getBukkitSoakContainers().sorted(Comparator.comparing(pl -> pl.metadata().id())).flatMap(pl -> SoakManager.<WrapperManager>getManager().getBukkitCommands(pl.getBukkitInstance()).stream()).filter(cmd -> {
+            if (cmd.getName().equalsIgnoreCase(name)) {
+                return true;
+            }
+            return cmd.getAliases().stream().anyMatch(alias -> alias.equalsIgnoreCase(name));
+        }).filter(cmd -> cmd instanceof PluginCommand).findFirst().map(cmd -> (PluginCommand) cmd).orElse(null);
         if (pluginCommand == null) {
             SoakManager.<WrapperManager>getManager().getLogger().warn("A Bukkit plugin attempted to access the command '" + name + "'. It however does not exist");
         }
@@ -652,15 +776,17 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public boolean dispatchCommand(@NotNull CommandSender sender, @NotNull String commandLine) throws CommandException {
-        throw NotImplementedException.createByLazy(SoakServer.class,
-                "dispatchCommand",
-                CommandSender.class,
-                String.class);
+        throw NotImplementedException.createByLazy(SoakServer.class, "dispatchCommand", CommandSender.class, String.class);
     }
 
     @Override
     public boolean addRecipe(@Nullable Recipe recipe) {
-        throw NotImplementedException.createByLazy(SoakServer.class, "addRecipe", Recipe.class);
+        return addRecipe(recipe, false);
+    }
+
+    @Override
+    public boolean addRecipe(@Nullable Recipe recipe, boolean resendRecipes) {
+        throw NotImplementedException.createByLazy(SoakServer.class, "addRecipe", Recipe.class, boolean.class);
     }
 
     @Override
@@ -675,21 +801,14 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @NotNull Iterator<Recipe> recipeIterator() {
-        return Sponge
-                .server()
-                .recipeManager()
-                .all()
-                .stream()
-                .<Recipe>map(recipe -> {
-                    try {
-                        return SoakRecipeMap.toBukkit(recipe);
-                    } catch (RuntimeException e) {
-                        e.printStackTrace();
-                        return null;
-                    }
-                })
-                .filter(Objects::nonNull)
-                .iterator();
+        return Sponge.server().recipeManager().all().stream().<Recipe>map(recipe -> {
+            try {
+                return SoakRecipeMap.toBukkit(recipe);
+            } catch (RuntimeException e) {
+                e.printStackTrace();
+                return null;
+            }
+        }).filter(Objects::nonNull).iterator();
     }
 
     @Override
@@ -713,10 +832,7 @@ public abstract class SoakServer implements SimpServer {
     }
 
     public ServerWorld defaultWorld() {
-        return this.spongeServer()
-                .worldManager()
-                .world(DefaultWorldKeys.DEFAULT)
-                .orElseThrow(() -> new RuntimeException("default world is not loaded"));
+        return this.spongeServer().worldManager().world(DefaultWorldKeys.DEFAULT).orElseThrow(() -> new RuntimeException("default world is not loaded"));
     }
 
     @Override
@@ -765,10 +881,7 @@ public abstract class SoakServer implements SimpServer {
         if (cached != null) {
             return cached;
         }
-        var future = Sponge.server()
-                .gameProfileManager()
-                .profile(name)
-                .thenCompose(profile -> Sponge.server().userManager().loadOrCreate(profile.uuid()));
+        var future = Sponge.server().gameProfileManager().profile(name).thenCompose(profile -> Sponge.server().userManager().loadOrCreate(profile.uuid()));
         return new SoakOfflinePlayer(future);
     }
 
@@ -777,10 +890,7 @@ public abstract class SoakServer implements SimpServer {
         if (Sponge.server().userManager().streamOfMatches(name).findAny().isEmpty()) {
             return null;
         }
-        var future = Sponge.server()
-                .userManager()
-                .load(name)
-                .thenApply(op -> op.orElseThrow(() -> new RuntimeException("Failed to load cached player")));
+        var future = Sponge.server().userManager().load(name).thenApply(op -> op.orElseThrow(() -> new RuntimeException("Failed to load cached player")));
         return new SoakOfflinePlayer(future);
     }
 
@@ -792,17 +902,24 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public org.bukkit.profile.@NotNull PlayerProfile createPlayerProfile(@Nullable UUID uuid, @Nullable String s) {
-        throw NotImplementedException.createByLazy(Server.class, "createPlayerProfile", UUID.class, String.class);
+        var opProfile = Sponge.server().gameProfileManager().cache().findById(uuid);
+        if (opProfile.isPresent()) {
+            return new SoakPlayerProfile(opProfile.get(), true);
+        }
+
+        var profile = GameProfile.of(uuid, s);
+        return new SoakPlayerProfile(profile, false);
+
     }
 
     @Override
-    public org.bukkit.profile.@NotNull PlayerProfile createPlayerProfile(@NotNull UUID uuid) {
-        throw NotImplementedException.createByLazy(Server.class, "createPlayerProfile", UUID.class);
+    public @NotNull org.bukkit.profile.PlayerProfile createPlayerProfile(@NotNull UUID uuid) {
+        return createPlayerProfile(uuid, null);
     }
 
     @Override
-    public org.bukkit.profile.@NotNull PlayerProfile createPlayerProfile(@NotNull String s) {
-        throw NotImplementedException.createByLazy(Server.class, "createPlayerProfile", String.class);
+    public @NotNull org.bukkit.profile.PlayerProfile createPlayerProfile(@NotNull String s) {
+        return createPlayerProfile(null, s);
     }
 
     @Override
@@ -843,25 +960,16 @@ public abstract class SoakServer implements SimpServer {
     public @NotNull Set<OfflinePlayer> getBannedPlayers() {
         var banService = Sponge.server().serviceProvider().banService();
         var userManager = Sponge.server().userManager();
-        CompletableFuture<Set<OfflinePlayer>> profiles = banService
-                .profileBans()
-                .thenCompose(collection -> {
-                    CompletableFuture<Optional<User>>[] futures = collection.stream().map(profile -> userManager.load(profile.profile())).toArray(CompletableFuture[]::new);
-                    return CompletableFuture
-                            .allOf(futures)
-                            .thenApply(v -> Arrays
-                                    .stream(futures)
-                                    .map(f -> {
-                                        try {
-                                            return f.get();
-                                        } catch (InterruptedException | ExecutionException e) {
-                                            throw new RuntimeException(e);
-                                        }
-                                    })
-                                    .filter(Optional::isPresent)
-                                    .map(opUser -> new SoakOfflinePlayer(opUser.get()))
-                                    .collect(Collectors.toSet()));
-                });
+        CompletableFuture<Set<OfflinePlayer>> profiles = banService.profileBans().thenCompose(collection -> {
+            CompletableFuture<Optional<User>>[] futures = collection.stream().map(profile -> userManager.load(profile.profile())).toArray(CompletableFuture[]::new);
+            return CompletableFuture.allOf(futures).thenApply(v -> Arrays.stream(futures).map(f -> {
+                try {
+                    return f.get();
+                } catch (InterruptedException | ExecutionException e) {
+                    throw new RuntimeException(e);
+                }
+            }).filter(Optional::isPresent).map(opUser -> new SoakOfflinePlayer(opUser.get())).collect(Collectors.toSet()));
+        });
 
         try {
             return profiles.get();
@@ -917,19 +1025,12 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public @NotNull Inventory createInventory(InventoryHolder arg0, @NotNull InventoryType arg1) {
-        throw NotImplementedException.createByLazy(Server.class,
-                "createInventory",
-                InventoryHolder.class,
-                InventoryType.class);
+        throw NotImplementedException.createByLazy(Server.class, "createInventory", InventoryHolder.class, InventoryType.class);
     }
 
     @Override
     public @NotNull Inventory createInventory(InventoryHolder arg0, @NotNull InventoryType arg1, @NotNull Component arg2) {
-        throw NotImplementedException.createByLazy(Server.class,
-                "createInventory",
-                InventoryHolder.class,
-                InventoryType.class,
-                Component.class);
+        throw NotImplementedException.createByLazy(Server.class, "createInventory", InventoryHolder.class, InventoryType.class, Component.class);
     }
 
     @Deprecated
@@ -951,7 +1052,7 @@ public abstract class SoakServer implements SimpServer {
     private Inventory createChestInventory(InventoryHolder holder, int size, @Nullable Component title) {
         int rows = (size / 9);
         var containerType = InventoryHelper.toChestContainerType(rows);
-        var plugin = GenericHelper.fromStackTrace();
+        var plugin = GeneralHelper.fromStackTrace();
         var inventory = ViewableInventory.builder().type(containerType).completeStructure().plugin(plugin).build();
         //TODO holder
         var bukkitInv = new SoakInventory<>(inventory);
@@ -1022,6 +1123,11 @@ public abstract class SoakServer implements SimpServer {
     }
 
     @Override
+    public void motd(@NotNull Component component) {
+        throw NotImplementedException.createByLazy(Server.class, "motd", Component.class);
+    }
+
+    @Override
     public @NotNull Component motd() {
         return this.spongeServer().motd();
     }
@@ -1030,6 +1136,11 @@ public abstract class SoakServer implements SimpServer {
     @Override
     public @NotNull String getMotd() {
         return LegacyComponentSerializer.legacySection().serialize(motd());
+    }
+
+    @Override
+    public void setMotd(@NotNull String s) {
+        motd(SoakMessageMap.toComponent(s));
     }
 
     @Override
@@ -1094,33 +1205,13 @@ public abstract class SoakServer implements SimpServer {
     }
 
     @Override
-    public ChunkGenerator.@NotNull ChunkData createVanillaChunkData(@NotNull World arg0, int arg1, int arg2) {
-        throw NotImplementedException.createByLazy(Server.class,
-                "createVanillaChunkData",
-                World.class,
-                int.class,
-                int.class);
-    }
-
-    @Override
     public @NotNull KeyedBossBar createBossBar(@NotNull NamespacedKey arg0, String arg1, @NotNull BarColor arg2, @NotNull BarStyle arg3, BarFlag[] arg4) {
-        throw NotImplementedException.createByLazy(Server.class,
-                "createBossBar",
-                NamespacedKey.class,
-                String.class,
-                BarColor.class,
-                BarStyle.class,
-                BarFlag[].class);
+        throw NotImplementedException.createByLazy(Server.class, "createBossBar", NamespacedKey.class, String.class, BarColor.class, BarStyle.class, BarFlag[].class);
     }
 
     @Override
     public @NotNull BossBar createBossBar(String arg0, @NotNull BarColor arg1, @NotNull BarStyle arg2, BarFlag[] arg3) {
-        throw NotImplementedException.createByLazy(Server.class,
-                "createBossBar",
-                String.class,
-                BarColor.class,
-                BarStyle.class,
-                BarFlag[].class);
+        throw NotImplementedException.createByLazy(Server.class, "createBossBar", String.class, BarColor.class, BarStyle.class, BarFlag[].class);
     }
 
     @Override
@@ -1185,11 +1276,6 @@ public abstract class SoakServer implements SimpServer {
     }
 
     @Override
-    public @NotNull BlockData createBlockData(@NotNull Material material, @Nullable Consumer<BlockData> consumer) {
-        throw NotImplementedException.createByLazy(Server.class, "createBlockData", Material.class, Consumer.class);
-    }
-
-    @Override
     public @NotNull BlockData createBlockData(@Nullable Material material, @Nullable String s) throws IllegalArgumentException {
         if (material == null && s == null) {
             throw new IllegalArgumentException("Both material and string cannot be null");
@@ -1201,15 +1287,19 @@ public abstract class SoakServer implements SimpServer {
             return createBlockData(material);
         }
         if (s.startsWith("[")) {
-            s = material.asBlock().orElseThrow(() -> new IllegalStateException("Item cannot be converted to BlockState")).key(RegistryTypes.BLOCK_TYPE).formatted() + s;
+            s = SoakBlockMap.toSponge(material).orElseThrow(() -> new IllegalStateException("Item cannot be converted to BlockState")).key(RegistryTypes.BLOCK_TYPE).formatted() + s;
         }
         var blockState = BlockState.fromString(s);
         return new SoakBlockData(blockState);
     }
 
     @Override
-    public @NotNull <T extends Keyed> Iterable<Tag<T>> getTags(@NotNull String registry, @NotNull Class<T> clazz) {
-        throw NotImplementedException.createByLazy(Server.class, "getTags", String.class, Class.class);
+    public @NotNull BlockData createBlockData(@NotNull Material material, @Nullable Consumer<? super BlockData> consumer) {
+        var result = createBlockData(material);
+        if (consumer != null) {
+            consumer.accept(result);
+        }
+        return result;
     }
 
     @Override
@@ -1273,8 +1363,7 @@ public abstract class SoakServer implements SimpServer {
         }
         try {
             //blocking -> need to fix this
-            return profileManager.uncached()
-                    .profile(id).thenApply(profile -> new SoakPlayerProfile(profile, false)).get();
+            return profileManager.uncached().profile(id).thenApply(profile -> new SoakPlayerProfile(profile, false)).get();
         } catch (InterruptedException | ExecutionException e) {
             throw new RuntimeException(e);
         }
@@ -1295,8 +1384,7 @@ public abstract class SoakServer implements SimpServer {
         }
         try {
             //blocking -> need to fix this
-            return profileManager.uncached()
-                    .profile(uuid).thenApply(profile -> new SoakPlayerProfile(name == null ? profile : profile.withName(name), false)).get();
+            return profileManager.uncached().profile(uuid).thenApply(profile -> new SoakPlayerProfile(name == null ? profile : profile.withName(name), false)).get();
         } catch (InterruptedException | ExecutionException e) {
             throw new RuntimeException(e);
         }
@@ -1316,8 +1404,7 @@ public abstract class SoakServer implements SimpServer {
         }
         try {
             //blocking -> need to fix this
-            return profileManager.uncached()
-                    .profile(name).thenApply(profile -> new SoakPlayerProfile(profile, false)).get();
+            return profileManager.uncached().profile(name).thenApply(profile -> new SoakPlayerProfile(profile, false)).get();
         } catch (InterruptedException | ExecutionException e) {
             throw new RuntimeException(e);
         }
@@ -1397,11 +1484,7 @@ public abstract class SoakServer implements SimpServer {
 
     @Override
     public void sendPluginMessage(@NotNull Plugin source, @NotNull String channel, byte[] message) {
-        throw NotImplementedException.createByLazy(Server.class,
-                "sendPluginMessage",
-                Plugin.class,
-                String.class,
-                byte[].class);
+        throw NotImplementedException.createByLazy(Server.class, "sendPluginMessage", Plugin.class, String.class, byte[].class);
     }
 
     @Override
