@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 
 public class NotImplementedException extends RuntimeException {
 
-    private Method interfaceMethod;
+    private final Method interfaceMethod;
 
     public NotImplementedException(Class<?> from, String methodName, Class<?>... parameters) throws NoSuchMethodException {
         this(from.getDeclaredMethod(methodName, parameters));
@@ -17,14 +17,29 @@ public class NotImplementedException extends RuntimeException {
         this.interfaceMethod = interfaceMethod;
     }
 
+    public static NotImplementedException createByLazy(Object fromObject, String name, Class<?>... parameters) {
+        Class<?> clazz = fromObject.getClass();
+        while (!clazz.equals(Object.class)) {
+            clazz = clazz.getSuperclass();
+            try {
+                return new NotImplementedException(clazz.getDeclaredMethod(name, parameters));
+            } catch (NoSuchMethodException e) {
+                continue;
+            }
+        }
+        //fail safe
+        throw new RuntimeException("The method is not implemented " + fromObject.getClass().getName() + "." + name + "(" + Arrays.stream(parameters).map(parameter -> parameter.getSimpleName() + " arg").collect(Collectors.joining(", ")) + ")");
+    }
+
     public static NotImplementedException createByLazy(Class<?> from, String name, Class<?>... parameters) {
         try {
             throw new NotImplementedException(from, name, parameters);
         } catch (NoSuchMethodException e) {
-            throw new RuntimeException("Could not create NotImplementedException", e);
+            throw new RuntimeException("Could not create NotImplementedException for " + from.getSimpleName() + "#" + name + "(" + Arrays.stream(parameters).map(parameter -> parameter.getSimpleName() + " arg").collect(Collectors.joining(", ")) + ")", e);
         }
     }
 
+    @Deprecated(forRemoval = true)
     public static NotImplementedException createByLazy(Class<?>... parameterTypes) {
         StackTraceElement stacktrace = Thread.currentThread().getStackTrace()[0];
         String className = stacktrace.getClassName();
