@@ -1,32 +1,38 @@
 package org.soak.plugin;
 
 import com.google.inject.Inject;
+import net.bytebuddy.dynamic.loading.ClassLoadingStrategy;
+import net.bytebuddy.jar.asm.MethodTooLargeException;
 import org.apache.logging.log4j.Logger;
 import org.apache.maven.artifact.versioning.ArtifactVersion;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.jetbrains.annotations.NotNull;
 import org.mosestream.MoseStream;
-import org.soak.Compatibility;
+import org.soak.NMSBounceLoader;
 import org.soak.WrapperManager;
 import org.soak.commands.soak.SoakCommand;
 import org.soak.config.SoakConfiguration;
 import org.soak.data.sponge.PortalCooldownCustomData;
 import org.soak.data.sponge.SoakKeys;
 import org.soak.fix.forge.ForgeFixCommons;
-import org.soak.hook.command.DynamicCommandListener;
-import org.soak.hook.command.SpongeDynamicCommand;
+import org.soak.generate.bukkit.AttributeTypeList;
+import org.soak.generate.bukkit.EntityTypeList;
+import org.soak.generate.bukkit.MaterialList;
 import org.soak.hook.event.HelpMapListener;
 import org.soak.io.SoakServerProperties;
 import org.soak.plugin.external.SoakConfig;
 import org.soak.plugin.loader.Locator;
 import org.soak.plugin.loader.common.AbstractSoakPluginContainer;
 import org.soak.plugin.loader.common.SoakPluginInjector;
+import org.soak.plugin.paper.loader.SoakPluginClassLoader;
+import org.soak.utils.RegisterUtils;
 import org.soak.utils.SoakMemoryStore;
 import org.soak.utils.log.CustomLoggerFormat;
 import org.soak.wrapper.SoakServer;
 import org.soak.wrapper.plugin.SoakPluginManager;
-import org.soak.wrapper.v1_19_R4.NMSBounceSoakServer;
+import org.soak.wrapper.v1_21_R2.NMSBounceSoakServer;
 import org.spongepowered.api.Server;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.command.Command;
@@ -39,28 +45,62 @@ import org.spongepowered.api.event.lifecycle.*;
 import org.spongepowered.api.item.inventory.ItemStack;
 import org.spongepowered.api.item.inventory.ItemStackSnapshot;
 import org.spongepowered.configurate.ConfigurateException;
+import org.spongepowered.configurate.yaml.YamlConfigurationLoader;
 import org.spongepowered.plugin.PluginContainer;
+import org.spongepowered.plugin.metadata.model.PluginDependency;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.lang.invoke.MethodHandles;
 import java.nio.file.Path;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.TreeSet;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.jar.JarFile;
 import java.util.logging.ConsoleHandler;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @org.spongepowered.plugin.builtin.jvm.Plugin("soak")
 public class SoakPlugin implements SoakExternalManager, WrapperManager {
 
     private static SoakPlugin plugin;
+    public final Collection<Class<?>> generatedClasses = new LinkedBlockingQueue<>();
     private final SoakConfiguration configuration;
     private final PluginContainer container;
     private final Logger logger;
-    private final Compatibility compatibility;
     private final SoakMemoryStore memoryStore = new SoakMemoryStore();
-
     private final SoakServerProperties serverProperties = new SoakServerProperties();
     private final ConsoleHandler consoleHandler = new ConsoleHandler();
+    private final int generatedClassesCount = 3;
+    private final Collection<SoakPluginContainer> loadedPlugins = new TreeSet<>((plugin, compare) -> {
+        var opDepends = plugin.metadata().dependencies().stream().filter(p -> p.id().equals(compare.metadata().id())).findAny();
+        if (opDepends.isPresent()) {
+            Integer result = switch (opDepends.get().loadOrder()) {
+                case PluginDependency.LoadOrder.AFTER -> 1;
+                case PluginDependency.LoadOrder.BEFORE -> -1;
+                default -> null;
+            };
+            if (result != null) {
+                return result;
+            }
+        }
+        opDepends = compare.metadata().dependencies().stream().filter(p -> p.id().equals(plugin.metadata().id())).findAny();
+        if (opDepends.isPresent()) {
+            Integer result = switch (opDepends.get().loadOrder()) {
+                case PluginDependency.LoadOrder.AFTER -> 1;
+                case PluginDependency.LoadOrder.BEFORE -> -1;
+                default -> null;
+            };
+            if (result != null) {
+                return result;
+            }
+        }
+        return 0;
+    });
 
     @Inject
     public SoakPlugin(PluginContainer pluginContainer, Logger logger) {
@@ -68,7 +108,6 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
         GlobalSoakData.MANAGER_INSTANCE = this;
         this.container = pluginContainer;
         this.logger = logger;
-        this.compatibility = new Compatibility();
         try {
             Path path = Sponge.configManager().pluginConfig(this.container).configPath();
             this.configuration = new SoakConfiguration(path.toFile());
@@ -87,14 +126,62 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
         return (SoakServer) Bukkit.getServer();
     }
 
+    private boolean hasDependency(Plugin plugin) {
+        String id = plugin.getName().toLowerCase();
+        return this
+                .loadedPlugins
+                .stream()
+                .anyMatch(pluginContainer -> pluginContainer
+                        .metadata()
+                        .dependencies()
+                        .stream()
+                        .anyMatch(dependency -> dependency.id().equals(id)));
+    }
+
+    @Listener
+    private void generateClasses(RegisterRegistryValueEvent.EngineScoped<Server> event) {
+        String creatingClass = "Material";
+        try {
+            var classLoader = SoakPlugin.class.getClassLoader();
+
+            var materialList = MaterialList.createMaterialList();
+            MaterialList.LOADED_CLASS = materialList.load(classLoader, ClassLoadingStrategy.Default.INJECTION).getLoaded();
+            generatedClasses.add(MaterialList.LOADED_CLASS);
+
+            creatingClass = "EntityType";
+            var entityTypeList = EntityTypeList.createEntityTypeList();
+            EntityTypeList.LOADED_CLASS = entityTypeList.load(classLoader, ClassLoadingStrategy.Default.INJECTION).getLoaded();
+            generatedClasses.add(EntityTypeList.LOADED_CLASS);
+
+            creatingClass = "Attribute";
+            var attributeList = AttributeTypeList.createEntityTypeList();
+            AttributeTypeList.LOADED_CLASS = attributeList.load(classLoader, ClassLoadingStrategy.Default.INJECTION).getLoaded();
+            generatedClasses.add(AttributeTypeList.LOADED_CLASS);
+        } catch (MethodTooLargeException e) {
+            throw new IllegalStateException("This is a problem with Bukkit's design: PaperMC seem to be making a fix with its hardfork: Too many entries in " + creatingClass, e);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
     @Override
     public Logger getLogger() {
         return this.logger;
     }
 
     @Override
-    public Stream<SoakPluginContainer> getBukkitContainers() {
+    public Collection<Class<?>> generatedClasses() {
+        return Collections.unmodifiableCollection(this.generatedClasses);
+    }
+
+    @Override
+    public Stream<SoakPluginContainer> getBukkitSoakContainers() {
         return getPlugins();
+    }
+
+    @Override
+    public Stream<PluginContainer> getBukkitPluginContainers() {
+        return getPlugins().map(SoakPluginContainer::getTrueContainer);
     }
 
     @Override
@@ -111,6 +198,15 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
         return this.container.metadata().version();
     }
 
+    @Override
+    public @NotNull SoakPluginClassLoader getSoakClassLoader(@NotNull SoakPluginContainer container) {
+        var classLoader = SoakPlugin.server().getSoakPluginManager().getContext(container.getBukkitInstance()).loader();
+        if (classLoader == null) {
+            throw new IllegalStateException("SoakPluginContainer has not been registered with a classloader");
+        }
+        return classLoader;
+    }
+
     public SoakMemoryStore getMemoryStore() {
         return this.memoryStore;
     }
@@ -122,21 +218,21 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
 
     @Override
     public Collection<org.bukkit.command.Command> getBukkitCommands(Plugin plugin) {
-        var pluginContainer = getContainer(plugin);
+        var pluginContainer = getSoakContainer(plugin);
         if (!(pluginContainer instanceof AbstractSoakPluginContainer aspc)) {
             throw new IllegalStateException("Plugin expended to be extending AbstractSoakPluginContainer");
         }
         return aspc.instance().commands();
     }
 
-    @Listener
-    public void registerCommands(RegisterCommandEvent<Command.Parameterized> event) {
-        event.register(this.container, SoakCommand.createSoakCommand(), "soak");
+    @Override
+    public boolean shouldMaterialListUseModded() {
+        return this.configuration.shouldMaterialListUseModded();
     }
 
     @Listener
-    public void registerFakeCommandLauncher(RegisterCommandEvent<Command.Raw> event) {
-        event.register(this.container, new SpongeDynamicCommand(), "soakdynamic");
+    public void registerCommands(RegisterCommandEvent<Command.Parameterized> event) {
+        event.register(this.container, SoakCommand.createSoakCommand(), "soak");
     }
 
     @Listener
@@ -157,9 +253,11 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
 
     @Listener(order = Order.FIRST)
     public void startingPlugin(StartingEngineEvent<Server> event) {
-        SoakRegister.startEnchantmentTypes(this.logger);
-        SoakRegister.startPotionEffects(this.logger);
+        //SoakRegister.startEnchantmentTypes(this.logger);
+        //SoakRegister.startPotionEffects(this.logger);
         PortalCooldownCustomData.createTickScheduler();
+
+        loadPlugins(true);
     }
 
     @Listener(order = Order.LAST)
@@ -211,8 +309,8 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
         thread.start();
     }
 
-    public Compatibility getCompatibility() {
-        return this.compatibility;
+    public boolean didClassesGenerate() {
+        return this.generatedClasses.size() == this.generatedClassesCount;
     }
 
     @Listener
@@ -226,15 +324,60 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
             }
         }
 
+        var loader = NMSBounceLoader.getLoader();
+        if (loader.hasNMSBounce()) {
+            try {
+                loader.extractNmsBounce();
+                loader.loadNMSBounce();
+            } catch (Throwable e) {
+                this.logger.warn("NMSBounce not active. Bukkit plugins wont be able to access NMS", e);
+            }
+        } else {
+            this.logger.warn("NMSBounce not active. Bukkit plugins wont be able to access NMS");
+        }
+
         this.consoleHandler.setFormatter(new CustomLoggerFormat());
 
         SoakServer server = new NMSBounceSoakServer(Sponge::server);
-        SoakPluginManager pluginManager = server.getSoakPluginManager();
         //noinspection deprecation
         Bukkit.setServer(server);
 
+        RegisterUtils.registerSerializable();
+
+        loadPlugins(false);
+        Sponge.eventManager().registerListeners(this.container, new HelpMapListener());
+    }
+
+    private void loadPlugins(boolean late) {
+        SoakPluginManager pluginManager = SoakPlugin.server().getSoakPluginManager();
+        var latePlugins = this.configuration.getLoadingLatePlugins();
         Collection<File> files = Locator.files();
         for (File file : files) {
+            try {
+                var jarFile = new JarFile(file);
+                var entry = jarFile.getJarEntry("plugin.yml");
+                if (entry == null) {
+                    entry = jarFile.getJarEntry("paper-plugin.yml");
+                }
+                if (entry == null) {
+                    continue;
+                }
+                var is = jarFile.getInputStream(entry);
+                var br = new BufferedReader(new InputStreamReader(is));
+                var yamlNode = YamlConfigurationLoader.builder().buildAndLoadString(br.lines().collect(Collectors.joining("\n")));
+                var pluginName = yamlNode.node("name").getString();
+                if (pluginName == null) {
+                    continue;
+                }
+                if ((late && !latePlugins.contains(pluginName)) || (!late && latePlugins.contains(pluginName))) {
+                    continue;
+                }
+                jarFile.close();
+            } catch (Throwable e) {
+                continue;
+            }
+
+
             JavaPlugin plugin;
             try {
                 //noinspection deprecation
@@ -248,21 +391,22 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
                 continue;
             }
 
-            SoakPluginContainer container = new AbstractSoakPluginContainer(file, plugin);
-            SoakPluginInjector.injectPlugin(container);
+            SoakPluginContainer container = new AbstractSoakPluginContainer(file, plugin, hasDependency(plugin) ? Order.EARLY : Order.DEFAULT);
+            container.downloadLibraries();
+            loadedPlugins.add(container);
             Sponge.eventManager().registerListeners(container, container.instance(), MethodHandles.lookup());
         }
+        SoakPluginInjector.injectPlugins(loadedPlugins);
+
         this.getPlugins().forEach(container -> ((AbstractSoakPluginContainer) container).instance().onPluginsConstructed());
-        Sponge.eventManager().registerListeners(this.container, new HelpMapListener());
-        Sponge.eventManager().registerListeners(this.container, new DynamicCommandListener());
+        if (late) {
+            this.getPlugins().forEach(container -> ((AbstractSoakPluginContainer) container).getBukkitInstance().onLoad());
+        }
     }
 
     public Stream<SoakPluginContainer> getPlugins() {
-        return Sponge.pluginManager()
-                .plugins()
-                .stream()
-                .filter(container -> container instanceof SoakPluginContainer)
-                .map(container -> (SoakPluginContainer) container);
+        return loadedPlugins
+                .stream();
     }
 
     public PluginContainer container() {
@@ -281,4 +425,5 @@ public class SoakPlugin implements SoakExternalManager, WrapperManager {
     public SoakConfig getConfig() {
         return this.configuration;
     }
+
 }
