@@ -4,12 +4,22 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
+import org.soak.plugin.SoakPlugin;
 import org.soak.plugin.SoakPluginContainer;
+import org.soak.plugin.paper.meta.SoakPluginMeta;
+import org.spongepowered.api.Sponge;
+import org.spongepowered.api.event.Order;
 import org.spongepowered.plugin.PluginContainer;
 import org.spongepowered.plugin.metadata.PluginMetadata;
 
 import java.io.File;
+import java.io.IOException;
+import java.net.MalformedURLException;
 import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.StandardCopyOption;
 import java.util.Optional;
 
 public class AbstractSoakPluginContainer implements SoakPluginContainer {
@@ -21,14 +31,14 @@ public class AbstractSoakPluginContainer implements SoakPluginContainer {
     private final JavaPlugin plugin;
     private final SoakPluginWrapper mainInstance;
 
-    public AbstractSoakPluginContainer(File bukkitPluginFile, JavaPlugin plugin) {
+    public AbstractSoakPluginContainer(File bukkitPluginFile, JavaPlugin plugin, Order order) {
         this.bukkitPluginFile = bukkitPluginFile;
         this.plugin = plugin;
         this.logger = LogManager.getLogger(plugin.getName());
         this.pluginMetadata = SoakPluginMetadata.fromPlugin(plugin);
 
         //temp
-        this.mainInstance = new SoakPluginWrapper(this);
+        this.mainInstance = new SoakPluginWrapper(this, order);
     }
 
     public File getPluginFile() {
@@ -61,6 +71,23 @@ public class AbstractSoakPluginContainer implements SoakPluginContainer {
     }
 
     @Override
+    public Optional<URI> locateResource(String path) {
+        boolean exists = this.plugin.getResource(path) != null;
+        if (!exists) {
+            return Optional.empty();
+        }
+
+        URI localPath = this.bukkitPluginFile.toURI();
+        URI ret = null;
+        try {
+            ret = localPath.relativize(new URI(path));
+        } catch (URISyntaxException e) {
+            throw new RuntimeException(e);
+        }
+        return Optional.of(ret);
+    }
+
+    /*@Override
     public Optional<URI> locateResource(URI relative) {
         boolean exists = this.plugin.getResource(relative.getPath()) != null;
         if (!exists) {
@@ -70,5 +97,5 @@ public class AbstractSoakPluginContainer implements SoakPluginContainer {
         URI localPath = this.bukkitPluginFile.toURI();
         URI ret = localPath.relativize(relative);
         return Optional.of(ret);
-    }
+    }*/
 }
