@@ -1,20 +1,23 @@
 package org.soak.plugin.loader.common;
 
 import org.soak.plugin.SoakPluginContainer;
-import org.soak.plugin.loader.forge.ForgePluginInjector;
 import org.soak.plugin.loader.vanilla.VanillaPluginInjector;
 
 import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.Collection;
 
 public interface SoakPluginInjector {
 
-    static void injectPlugin(SoakPluginContainer container) {
+    static void injectPlugins(Collection<SoakPluginContainer> containers) {
         try {
-            VanillaPluginInjector.injectPluginToPlatform(container);
+            for (var container : containers) {
+                VanillaPluginInjector.injectPluginToPlatform(container);
+            }
         } catch (NoSuchMethodException e) {
-            //load forge
+            //load neo forge
             try {
-                ForgePluginInjector.injectPluginToPlatform(container);
+                injectPluginToPlatform("org.soak.plugin.loader.neo.NeoPluginInjector", containers);
             } catch (Throwable ex) {
                 throw new RuntimeException(ex);
             }
@@ -23,31 +26,22 @@ public interface SoakPluginInjector {
         }
     }
 
+    private static void injectPluginToPlatform(String className, Collection<SoakPluginContainer> containers) throws ClassNotFoundException, NoSuchMethodException, InvocationTargetException, IllegalAccessException {
+        var method = Class.forName(className).getDeclaredMethod("injectPluginToPlatform", Collection.class);
+        method.invoke(null, containers);
+    }
+
     static void removePluginFromPlatform(String id) {
         try {
             VanillaPluginInjector.removePluginFromPlatform(id);
-        } catch (NoSuchFieldException e) {
-            try {
-                ForgePluginInjector.removePluginFromPlatform(id);
-            } catch (Throwable ex) {
-                throw new RuntimeException(ex);
-            }
-        } catch (IllegalAccessException e) {
-            throw new RuntimeException(e);
+        } catch (IllegalAccessException | NoSuchFieldException e) {
         }
     }
 
     static void removePluginFromPlatform(SoakPluginContainer container) {
         try {
             VanillaPluginInjector.removePluginFromPlatform(container);
-        } catch (NoSuchFieldException e) {
-            try {
-                ForgePluginInjector.removePluginFromPlatform(container);
-            } catch (Throwable ex) {
-                throw new RuntimeException(ex);
-            }
-        } catch (IllegalAccessException e) {
-            throw new RuntimeException(e);
+        } catch (IllegalAccessException | NoSuchFieldException e) {
         }
     }
 
