@@ -1,5 +1,6 @@
 package org.soak.plugin.loader.common;
 
+import io.leangen.geantyref.TypeToken;
 import org.bukkit.command.PluginCommandYamlParser;
 import org.bukkit.plugin.Plugin;
 import org.soak.command.BukkitRawCommand;
@@ -7,8 +8,12 @@ import org.soak.plugin.SoakManager;
 import org.soak.plugin.SoakPlugin;
 import org.soak.plugin.SoakPluginContainer;
 import org.spongepowered.api.Server;
+import org.spongepowered.api.Sponge;
 import org.spongepowered.api.command.Command;
+import org.spongepowered.api.command.manager.CommandFailedRegistrationException;
+import org.spongepowered.api.event.EventListenerRegistration;
 import org.spongepowered.api.event.Listener;
+import org.spongepowered.api.event.Order;
 import org.spongepowered.api.event.lifecycle.RegisterCommandEvent;
 import org.spongepowered.api.event.lifecycle.StartedEngineEvent;
 import org.spongepowered.api.event.lifecycle.StartingEngineEvent;
@@ -24,8 +29,9 @@ public class SoakPluginWrapper {
     private final SoakPluginContainer pluginContainer;
     private final Collection<org.bukkit.command.Command> commands = new LinkedHashSet<>();
     private boolean hasRunShutdown;
+    private boolean loadedCommandsEarly;
 
-    public SoakPluginWrapper(SoakPluginContainer pluginContainer) {
+    public SoakPluginWrapper(SoakPluginContainer pluginContainer, Order order) {
         this.pluginContainer = pluginContainer;
 
         Plugin plugin = this.pluginContainer.getBukkitInstance();
@@ -39,6 +45,8 @@ public class SoakPluginWrapper {
                 SoakManager.getManager().displayError(e, plugin);
             }
         }
+        Sponge.eventManager().registerListener(EventListenerRegistration.builder(new TypeToken<StartedEngineEvent<Server>>() {
+        }).order(order).listener(this::onPluginEnable).plugin(pluginContainer.getTrueContainer()).build());
     }
 
     public void onPluginsConstructed() {
@@ -49,6 +57,8 @@ public class SoakPluginWrapper {
         Plugin plugin = this.pluginContainer.getBukkitInstance();
         try {
             plugin.onEnable();
+            SoakPlugin.plugin().logger().info(this.pluginContainer.metadata().id() + " loaded early");
+
         } catch (Throwable e) {
             SoakManager.getManager().displayError(e, plugin);
         }
@@ -64,14 +74,15 @@ public class SoakPluginWrapper {
 
     @Listener
     public void onCommandRegister(RegisterCommandEvent<Command.Raw> event) {
-        this.commands.forEach(cmd -> event.register(this.pluginContainer,
-                new BukkitRawCommand(this.pluginContainer, cmd),
-                cmd.getName(),
-                cmd.getAliases().toArray(String[]::new)));
+        loadedCommandsEarly = true;
+        this.commands.forEach(cmd -> event.register(this.pluginContainer, new BukkitRawCommand(this.pluginContainer, cmd), cmd.getName(), cmd.getAliases().toArray(String[]::new)));
     }
 
     @Listener
     public void onPluginLoad(StartingEngineEvent<Server> event) {
+        if (!SoakPlugin.plugin().didClassesGenerate()) {
+            return;
+        }
         var earlyPlugins = SoakPlugin.plugin().config().getLoadingEarlyPlugins();
         if (earlyPlugins.contains(this.pluginContainer.getBukkitInstance().getName())) {
             return;
@@ -87,15 +98,34 @@ public class SoakPluginWrapper {
     //issue
     //Bukkit plugins assume everything is loaded when onEnable is run, this is because Craftbukkit loads everything before onEnable is used ....
     //using StartedEngineEvent despite the timing known to be incorrect
-    @Listener
     public void onPluginEnable(StartedEngineEvent<Server> event) {
+        if (!SoakPlugin.plugin().didClassesGenerate()) {
+            return;
+        }
         var earlyPlugins = SoakPlugin.plugin().config().getLoadingEarlyPlugins();
         if (earlyPlugins.contains(this.pluginContainer.getBukkitInstance().getName())) {
             return;
         }
         Plugin plugin = this.pluginContainer.getBukkitInstance();
+        if(!loadedCommandsEarly) {
+            //load commands
+            var registerCmd = Sponge.server().commandManager().registrar(Command.Raw.class).orElseThrow(() -> new IllegalStateException("Cannot load the command raw register for " + pluginContainer.metadata().id()));
+            this.commands.forEach(cmd -> {
+                try {
+                    registerCmd.register(this.pluginContainer, new BukkitRawCommand(this.pluginContainer, cmd), cmd.getName(), cmd.getAliases().toArray(String[]::new));
+                } catch (CommandFailedRegistrationException ex) {
+                    try {
+                        registerCmd.register(this.pluginContainer, new BukkitRawCommand(this.pluginContainer, cmd), cmd.getName());
+                    } catch (CommandFailedRegistrationException ex1) {
+                        pluginContainer.logger().warn("Cannot register command, skipping: ", ex1);
+                    }
+                }
+            });
+        }
+
         try {
             plugin.onEnable();
+            SoakPlugin.plugin().logger().info(this.pluginContainer.metadata().id() + " loaded late");
         } catch (Throwable e) {
             SoakManager.getManager().displayError(e, plugin);
         }
@@ -103,6 +133,9 @@ public class SoakPluginWrapper {
 
     @Listener
     public void onPluginDisable(StoppingEngineEvent<Server> event) {
+        if (!SoakPlugin.plugin().didClassesGenerate()) {
+            return;
+        }
         if (hasRunShutdown) {
             return;
         }
