@@ -2,7 +2,6 @@ package org.soak.plugin.loader.vanilla;
 
 import org.jetbrains.annotations.NotNull;
 import org.soak.plugin.SoakPluginContainer;
-import org.soak.plugin.loader.common.SoakPluginInjector;
 import org.soak.utils.ReflectionHelper;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.plugin.PluginManager;
@@ -12,7 +11,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import java.util.Map;
 
-public class VanillaPluginInjector implements SoakPluginInjector {
+public class VanillaPluginInjector {
     public static void injectPluginToPlatform(SoakPluginContainer container) throws InvocationTargetException, IllegalAccessException, NoSuchMethodException {
         PluginManager manager = Sponge.pluginManager();
         ReflectionHelper.runMethod(manager, "addPlugin", container);
