@@ -14,26 +14,29 @@ import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.CheckReturnValue;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.soak.Compatibility;
 import org.soak.WrapperManager;
 import org.soak.io.SoakServerProperties;
+import org.soak.plugin.paper.loader.FoundClassLoader;
 import org.soak.plugin.paper.loader.SoakPluginClassLoader;
 import org.soak.plugin.paper.meta.SoakPluginMetaBuilder;
 import org.soak.utils.SoakMemoryStore;
-import org.soak.wrapper.v1_19_R4.NMSBounceSoakServer;
+import org.soak.wrapper.v1_21_R2.NMSBounceSoakServer;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.plugin.PluginContainer;
 
 import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
-import java.net.*;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.LinkedTransferQueue;
 import java.util.function.Function;
 import java.util.logging.ConsoleHandler;
@@ -46,33 +49,34 @@ public class AbstractSpongePluginMain implements SoakInternalManager, WrapperMan
     final Collection<Command> commands = new LinkedTransferQueue<>();
     final InternalSoakPluginContainer soakPluginContainer = new InternalSoakPluginContainer(this);
     final PluginContainer container;
-    private final Function<ClassLoader, Class<? extends JavaPlugin>> loaderToMain;
+    private final Function<FoundClassLoader, Class<? extends JavaPlugin>> loaderToMain;
+    private final Function<Class<? extends JavaPlugin>, FoundClassLoader> mainToLoader;
     private final String[] pathToPlugins;
     private final Logger logger;
-    private final Compatibility compatibility;
     private final SoakServerProperties serverProperties = new SoakServerProperties();
     private final ConsoleHandler consoleHandler = new ConsoleHandler();
     private final SoakMemoryStore memoryStore = new SoakMemoryStore();
+    private final Collection<Class<?>> generatedClasses = new LinkedBlockingQueue<>();
     @Nullable
-    private URLClassLoader loader;
+    private FoundClassLoader loader;
 
     @Nullable
     private JavaPlugin plugin;
 
     @Deprecated
-    public AbstractSpongePluginMain(Function<ClassLoader, Class<? extends JavaPlugin>> loaderToMain, Logger logger, PluginContainer container) {
-        this(loaderToMain, logger, container, new String[0]);
+    public AbstractSpongePluginMain(Function<Class<? extends JavaPlugin>, FoundClassLoader> mainToLoader, Function<FoundClassLoader, Class<? extends JavaPlugin>> loaderToMain, Logger logger, PluginContainer container) {
+        this(mainToLoader, loaderToMain, logger, container, new String[0]);
     }
 
-    public AbstractSpongePluginMain(Function<ClassLoader, Class<? extends JavaPlugin>> loaderToMain, Logger logger, PluginContainer container, String... pathsToPlugins) {
+    public AbstractSpongePluginMain(Function<Class<? extends JavaPlugin>, FoundClassLoader> mainToLoader, Function<FoundClassLoader, Class<? extends JavaPlugin>> loaderToMain, Logger logger, PluginContainer container, String... pathsToPlugins) {
         if (pathsToPlugins.length == 0) {
             throw new IllegalStateException("'PathsToPlugins' needs to be filled");
         }
+        this.mainToLoader = mainToLoader;
         this.loaderToMain = loaderToMain;
         this.pathToPlugins = pathsToPlugins;
         this.logger = logger;
         this.container = container;
-        this.compatibility = new Compatibility();
         GlobalSoakData.MANAGER_INSTANCE = this;
         Bukkit.setServer(new NMSBounceSoakServer(Sponge::server));
         Sponge.eventManager().registerListeners(container, new SoakWrapperListener(this));
@@ -93,21 +97,18 @@ public class AbstractSpongePluginMain implements SoakInternalManager, WrapperMan
         List<File> copiedFiles = new LinkedList<>();
         for (var path : this.pathToPlugins) {
             try {
-                var internalPath = new URI(path);
-                var inputStream = this.container.openResource(internalPath).orElseThrow(() -> new RuntimeException("Could not find '" + path + "'"));
+                var inputStream = this.container.openResource(path).orElseThrow(() -> new RuntimeException("Could not find '" + path + "'"));
                 var location = new File("temp/soak/" + path);
                 location.deleteOnExit();
                 location.getParentFile().mkdirs();
                 Files.copy(inputStream, location.toPath(), StandardCopyOption.REPLACE_EXISTING);
                 copiedFiles.add(location);
-            } catch (URISyntaxException e) {
-                throw new RuntimeException(e);
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
         }
 
-        loader = new URLClassLoader(copiedFiles.stream().map(file -> {
+        loader = new SoakClassLoader(copiedFiles.stream().map(file -> {
             try {
                 var url = file.toURI().toURL();
                 this.container.logger().info("Loading Bukkit file of '" + url.toString() + "'");
@@ -144,7 +145,7 @@ public class AbstractSpongePluginMain implements SoakInternalManager, WrapperMan
 
         Class<? extends JavaPlugin> javaPluginClass = this.loaderToMain.apply(loader);
         plugin = javaPluginClass.getConstructor().newInstance();
-        SoakPluginClassLoader.setupPlugin(plugin, this.container.metadata().id(), getPluginFile(), new File(getPluginFolder(), "config.yml"), getPluginFolder(), this::getPluginMeta, () -> loader);
+        SoakPluginClassLoader.setupPlugin(plugin, this.container.metadata().id(), getPluginFile(), new File(getPluginFolder(), "config.yml"), getPluginFolder(), this::getPluginMeta, () -> new SoakPluginMetaBuilder().from(this.getPluginMeta()).build(), () -> (ClassLoader) loader);
         this.commands.addAll(PluginCommandYamlParser.parse(plugin));
     }
 
@@ -199,8 +200,18 @@ public class AbstractSpongePluginMain implements SoakInternalManager, WrapperMan
     }
 
     @Override
-    public Stream<SoakPluginContainer> getBukkitContainers() {
+    public Collection<Class<?>> generatedClasses() {
+        return Collections.unmodifiableCollection(this.generatedClasses);
+    }
+
+    @Override
+    public Stream<SoakPluginContainer> getBukkitSoakContainers() {
         return Stream.of(this.soakPluginContainer);
+    }
+
+    @Override
+    public Stream<PluginContainer> getBukkitPluginContainers() {
+        return getBukkitSoakContainers().map(SoakPluginContainer::getTrueContainer);
     }
 
     @Override
@@ -232,6 +243,11 @@ public class AbstractSpongePluginMain implements SoakInternalManager, WrapperMan
     }
 
     @Override
+    public boolean shouldMaterialListUseModded() {
+        return false;
+    }
+
+    @Override
     public ConsoleHandler getConsole() {
         return this.consoleHandler;
     }
@@ -242,7 +258,7 @@ public class AbstractSpongePluginMain implements SoakInternalManager, WrapperMan
     }
 
     @Override
-    public Compatibility getCompatibility() {
-        return compatibility;
+    public @NotNull FoundClassLoader getSoakClassLoader(@NotNull SoakPluginContainer container) {
+        return this.mainToLoader.apply(container.getBukkitInstance().getClass());
     }
 }
