@@ -89,8 +89,8 @@ public class SoakWrapperListener {
 
     @Listener(order = Order.FIRST)
     public void startingPlugin(StartingEngineEvent<Server> event) {
-        SoakRegister.startEnchantmentTypes(pluginMain.getLogger());
-        SoakRegister.startPotionEffects(pluginMain.getLogger());
+        //SoakRegister.startEnchantmentTypes(pluginMain.getLogger());
+        //SoakRegister.startPotionEffects(pluginMain.getLogger());
         PortalCooldownCustomData.createTickScheduler();
     }
 
