@@ -33,7 +33,7 @@ public class SoakSpawnReasonMap {
         throw new RuntimeException("Unknown SpawnType of " + type.key(RegistryTypes.SPAWN_TYPE).asString());
     }
 
-    public static CreatureSpawnEvent.SpawnReason toBukkit(SpawnType type, Cause cause){
+    public static CreatureSpawnEvent.SpawnReason toBukkit(SpawnType type, Cause cause) {
         return toBukkit(type)
                 .entrySet()
                 .stream()
@@ -45,14 +45,15 @@ public class SoakSpawnReasonMap {
 
     public static DefaultedRegistryReference<SpawnType> toSponge(CreatureSpawnEvent.SpawnReason reason) {
         return switch (reason) {
-            case BEEHIVE, SPAWNER -> SpawnTypes.MOB_SPAWNER;
+            case BEEHIVE, SPAWNER, TRIAL_SPAWNER -> SpawnTypes.MOB_SPAWNER;
             case EGG, SPAWNER_EGG -> SpawnTypes.SPAWN_EGG;
             case BREEDING, OCELOT_BABY -> SpawnTypes.BREEDING;
             case BUILD_IRONGOLEM, BUILD_SNOWMAN, BUILD_WITHER, SILVERFISH_BLOCK -> SpawnTypes.BLOCK_SPAWNING;
-            case COMMAND, MOUNT, CURED, TRAP, RAID, CUSTOM, DROWNED, DUPLICATION, ENDER_PEARL, FROZEN, INFECTION, JOCKEY, LIGHTNING, METAMORPHOSIS, NATURAL, NETHER_PORTAL, PATROL, PIGLIN_ZOMBIFIED, REINFORCEMENTS, SHEARED, SHOULDER_ENTITY, SLIME_SPLIT, SPELL, VILLAGE_DEFENSE, VILLAGE_INVASION ->
-                    SpawnTypes.CUSTOM;
+            case ENCHANTMENT, OMINOUS_ITEM_SPAWNER, DEFAULT, POTION_EFFECT -> SpawnTypes.WORLD_SPAWNER;
+            case COMMAND, MOUNT, CURED, TRAP, RAID, CUSTOM, DROWNED, DUPLICATION, ENDER_PEARL, FROZEN, INFECTION,
+                 JOCKEY, LIGHTNING, METAMORPHOSIS, NATURAL, NETHER_PORTAL, PATROL, PIGLIN_ZOMBIFIED, REINFORCEMENTS,
+                 SHEARED, SHOULDER_ENTITY, SLIME_SPLIT, SPELL, VILLAGE_DEFENSE, VILLAGE_INVASION -> SpawnTypes.CUSTOM;
             case DISPENSE_EGG -> SpawnTypes.DISPENSE;
-            case DEFAULT -> SpawnTypes.WORLD_SPAWNER;
             case EXPLOSION -> SpawnTypes.TNT_IGNITE;
             case CHUNK_GEN -> SpawnTypes.CHUNK_LOAD;
         };
