@@ -24,7 +24,7 @@ public class SoakItemFlagMap {
             return ItemFlag.HIDE_PLACED_ON;
         }
         if (Keys.HIDE_MISCELLANEOUS.equals(key)) {
-            return ItemFlag.HIDE_POTION_EFFECTS;
+            return ItemFlag.HIDE_ADDITIONAL_TOOLTIP;
         }
         throw new RuntimeException("Unknown mapping from sponge key to ItemFlag of " + key.key().formatted());
     }
@@ -32,7 +32,7 @@ public class SoakItemFlagMap {
     public static Key<Value<Boolean>> toSponge(ItemFlag flag) {
         switch (flag) {
             case HIDE_ENCHANTS:
-                return Keys.HIDE_ENCHANTMENTS;
+                return Keys.HIDE_STORED_ENCHANTMENTS;
             case HIDE_ATTRIBUTES:
                 return Keys.HIDE_ATTRIBUTES;
             case HIDE_UNBREAKABLE:
@@ -41,10 +41,10 @@ public class SoakItemFlagMap {
                 return Keys.HIDE_CAN_DESTROY;
             case HIDE_PLACED_ON:
                 return Keys.HIDE_CAN_PLACE;
-            case HIDE_POTION_EFFECTS:
+            case HIDE_ADDITIONAL_TOOLTIP:
                 return Keys.HIDE_MISCELLANEOUS;
             case HIDE_DYE:
-                throw new IllegalStateException("Unknown key map for HIDE_DYE on ItemFlag");
+                return Keys.HIDE_ENCHANTMENTS;
             default:
                 throw new RuntimeException("Unknown ItemFlag of " + flag.name());
         }
