@@ -5,20 +5,20 @@ import org.bukkit.command.Command;
 import org.bukkit.command.SimpleCommandMap;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.soak.WrapperManager;
 import org.soak.command.BukkitRawCommand;
 import org.soak.plugin.SoakManager;
-import org.soak.utils.GenericHelper;
-import org.soak.wrapper.plugin.SoakPluginManager;
+import org.soak.utils.GeneralHelper;
 import org.spongepowered.api.Sponge;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.Objects;
 
 public class SoakCommandMap extends SimpleCommandMap {
 
 
     public SoakCommandMap() {
-        super(Bukkit.getServer());
+        super(Bukkit.getServer(), new HashMap<>());
     }
 
     @Override
@@ -31,8 +31,8 @@ public class SoakCommandMap extends SimpleCommandMap {
         boolean result = command.register(this);
 
         var rawRegister = Sponge.server().commandManager().registrar(org.spongepowered.api.command.Command.Raw.class).orElseThrow(() -> new RuntimeException("Cannot register late command of '" + label + "'"));
-        var plugin = GenericHelper.fromStackTrace();
-        var soakPlugin = SoakManager.getManager().getContainer(plugin).orElseThrow(() -> new RuntimeException("Cannot get the soakPlugin from '" + plugin.metadata().id() + "'"));
+        var plugin = GeneralHelper.fromStackTrace();
+        var soakPlugin = SoakManager.getManager().getSoakContainer(plugin).orElseThrow(() -> new RuntimeException("Cannot get the soakPlugin from '" + plugin.metadata().id() + "'"));
         var bukkitRawWrapper = new BukkitRawCommand(soakPlugin, command);
         rawRegister.register(plugin, bukkitRawWrapper, command.getName(), command.getAliases().toArray(String[]::new));
         return result;
