@@ -11,54 +11,95 @@ import org.bukkit.attribute.AttributeModifier;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemFlag;
+import org.bukkit.inventory.ItemRarity;
 import org.bukkit.inventory.meta.Damageable;
 import org.bukkit.inventory.meta.ItemMeta;
+import org.bukkit.inventory.meta.components.FoodComponent;
+import org.bukkit.inventory.meta.components.JukeboxPlayableComponent;
+import org.bukkit.inventory.meta.components.ToolComponent;
 import org.bukkit.inventory.meta.tags.CustomItemTagContainer;
 import org.bukkit.persistence.PersistentDataContainer;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.mose.collection.stream.builder.CollectionStreamBuilder;
 import org.soak.exception.NotImplementedException;
+import org.soak.map.SoakAttributeMap;
+import org.soak.map.SoakBlockMap;
 import org.soak.map.SoakMessageMap;
 import org.soak.map.item.SoakEnchantmentTypeMap;
 import org.soak.map.item.SoakItemFlagMap;
-import org.soak.utils.DataHelper;
+import org.soak.map.item.SoakItemStackMap;
+import org.soak.map.item.inventory.SoakEquipmentMap;
+import org.soak.wrapper.inventory.meta.food.SoakFoodComponent;
+import org.soak.wrapper.inventory.meta.tool.SoakToolComponent;
 import org.soak.wrapper.persistence.SoakImmutablePersistentDataContainer;
 import org.soak.wrapper.persistence.SoakMutablePersistentDataContainer;
 import org.spongepowered.api.data.DataHolder;
 import org.spongepowered.api.data.Key;
 import org.spongepowered.api.data.Keys;
+import org.spongepowered.api.data.persistence.DataView;
 import org.spongepowered.api.data.value.ListValue;
 import org.spongepowered.api.data.value.SetValue;
 import org.spongepowered.api.data.value.Value;
-import org.spongepowered.api.data.value.ValueContainer;
-import org.spongepowered.api.entity.Entity;
-import org.spongepowered.api.item.ItemType;
+import org.spongepowered.api.entity.attribute.type.AttributeType;
 import org.spongepowered.api.item.enchantment.EnchantmentType;
 import org.spongepowered.api.item.inventory.ItemStack;
+import org.spongepowered.api.item.inventory.ItemStackLike;
 import org.spongepowered.api.item.inventory.ItemStackSnapshot;
+import org.spongepowered.api.item.inventory.equipment.EquipmentType;
+import org.spongepowered.api.registry.RegistryTypes;
+import org.spongepowered.api.util.Ticks;
 
 import java.util.*;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public abstract class AbstractItemMeta implements ItemMeta, Damageable {
 
-    protected ValueContainer container;
+    protected ItemStackLike container;
 
-    protected AbstractItemMeta(ValueContainer container) {
-        if (!(container instanceof ItemStack || container instanceof ItemStackSnapshot || container instanceof Entity)) {
-            throw new RuntimeException("Must be either a ItemStack or ItemStackSnapshot");
-        }
+    protected AbstractItemMeta(ItemStackLike container) {
         this.container = container;
     }
 
-    public void manipulate(Function<ValueContainer, ValueContainer> function) {
-        this.container = function.apply(this.container);
+    /*
+    Required for ConfigurationSerializable
+     */
+    public static ItemMeta deserialize(Map<String, Object> values) {
+        return ItemMetaSerializer.deserialize(values);
     }
 
     public boolean isSnapshot() {
         return this.container instanceof ItemStackSnapshot;
+    }
+
+    public ItemStackLike sponge() {
+        return this.container;
+    }
+
+    public int durability() {
+        return this.container.get(Keys.ITEM_DURABILITY).orElse(0);
+    }
+
+    public void setDurability(int durability) {
+        set(Keys.ITEM_DURABILITY, durability);
+    }
+
+    public int maxDurability() {
+        return this.container.get(Keys.MAX_DURABILITY).orElse(0);
+    }
+
+    public int quantity() {
+        return container.quantity();
+    }
+
+    public void setQuantity(int quantity) {
+        if (this.container instanceof ItemStackSnapshot snapshot) {
+            this.container = snapshot.asMutableCopy();
+        }
+        ((ItemStack) this.container).setQuantity(quantity);
     }
 
     public void copyInto(ItemMeta meta) {
@@ -66,49 +107,29 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
             throw new RuntimeException("ItemMeta must implement AbstractItemMeta");
         }
         var into = (AbstractItemMeta) meta;
-        if (into.container instanceof ItemStack) {
-            this.container = DataHelper.copyInto((ItemStack) this.container,
-                    this.asSnapshot().orElseThrow(() -> new RuntimeException("Failed check")));
-            return;
-        }
-        if (into.container instanceof Entity) {
-            this.container = DataHelper.copyInto((Entity) into.container, (DataHolder) this.container);
-        }
-        this.container = DataHelper.copyInto((ItemStackSnapshot) into.container,
-                this.asSnapshot().orElseThrow(() -> new RuntimeException("Failed checks")));
+        into.container = this.container.copy();
     }
 
-    public Optional<ItemStack> asStack() {
-        if (this.container instanceof ItemStack) {
-            return Optional.of((ItemStack) this.container);
-        }
-        if (this.container instanceof ItemStackSnapshot) {
-            return Optional.of(((ItemStackSnapshot) this.container).createStack());
-        }
-        return Optional.empty();
+    public ItemStack asStack() {
+        return this.container.asMutableCopy();
     }
 
-    public Optional<ItemStackSnapshot> asSnapshot() {
-        if (this.container instanceof ItemStack) {
-            return Optional.of(((ItemStack) this.container).createSnapshot());
-        }
-        if (this.container instanceof ItemStackSnapshot) {
-            return Optional.of((ItemStackSnapshot) this.container);
-        }
-        return Optional.empty();
+    public ItemStackSnapshot asSnapshot() {
+        return this.container.asImmutable();
     }
 
-    protected DataHolder.Immutable<?> copyToImmutable() {
-        if (this.container instanceof ItemStack) {
-            return ((ItemStack) this.container).createSnapshot();
-        }
-        if (this.container instanceof ItemStackSnapshot) {
-            return ((ItemStackSnapshot) this.container).copy();
-        }
-        return ((Entity) this.container).createSnapshot();
+    public boolean addAttribute(AttributeType type, org.spongepowered.api.entity.attribute.AttributeModifier modifier, Stream<EquipmentType> equipmentTypes) {
+        var newStack = this.container.asMutableCopy();
+        AtomicBoolean hasApplied = new AtomicBoolean();
+        equipmentTypes.forEach(equipmentType -> {
+            hasApplied.set(true);
+            newStack.addAttributeModifier(type, modifier, equipmentType);
+        });
+        this.container = newStack;
+        return hasApplied.get();
     }
 
-    protected <T> void set(@NotNull Key<Value<T>> key, @Nullable T value) throws RuntimeException {
+    public <T> void set(@NotNull Key<Value<T>> key, @Nullable T value) throws RuntimeException {
         if (value == null) {
             remove(key);
             return;
@@ -117,16 +138,16 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
             ((DataHolder.Mutable) this.container).offer(key, value);
             return;
         }
-        var opStack = ((DataHolder.Immutable) this.container).with(key, value);
+        var opStack = ((ItemStackSnapshot) this.container).with(key, value);
         if (opStack.isEmpty()) {
             throw new RuntimeException("Key of " + key.key()
                     .formatted() + " is not supported with ItemStackSnapshot");
         }
 
-        this.container = (ValueContainer) opStack.get();
+        this.container = opStack.get();
     }
 
-    protected <T> void setList(@NotNull Key<ListValue<T>> key, @Nullable List<T> value) {
+    public <T> void setList(@NotNull Key<ListValue<T>> key, @Nullable List<T> value) {
         if (value == null) {
             remove(key);
             return;
@@ -135,7 +156,7 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
             ((DataHolder.Mutable) this.container).offer(key, value);
             return;
         }
-        this.container = ((DataHolder.Immutable<?>) this.container).with(key, value)
+        this.container = ((ItemStackSnapshot) this.container).with(key, value)
                 .orElseThrow(() -> new RuntimeException("Key of " + key.key()
                         .formatted() + " is not supported with ItemStackSnapshot"));
     }
@@ -149,16 +170,9 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
             ((DataHolder.Mutable) this.container).offer(key, value);
             return;
         }
-        this.container = ((DataHolder.Immutable<?>) this.container).with(key, value)
+        this.container = ((ItemStackSnapshot) this.container).with(key, value)
                 .orElseThrow(() -> new RuntimeException("Key of " + key.key()
                         .formatted() + " is not supported with ItemStackSnapshot"));
-    }
-
-    private ItemType type() {
-        if (this.container instanceof ItemStack) {
-            return ((ItemStack) this.container).type();
-        }
-        return ((ItemStackSnapshot) this.container).type();
     }
 
     protected void remove(@NotNull Key<?> key) {
@@ -190,13 +204,14 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
     public @NotNull String getDisplayName() {
         Component displayName = displayName();
         if (displayName == null) {
-            displayName = this.type().asComponent();
+            displayName = this.container.type().asComponent();
         }
         return SoakMessageMap.mapToBukkit(displayName);
     }
 
     @Override
     public void setDisplayName(@Nullable String name) {
+        this.container = this.container.asMutableCopy(); //fixed ItemStackSnapshot not supporting the key
         if (name == null) {
             remove(Keys.CUSTOM_NAME);
             return;
@@ -355,12 +370,13 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
     private boolean modifyEnchantments(Function<List<org.spongepowered.api.item.enchantment.Enchantment>, List<org.spongepowered.api.item.enchantment.Enchantment>> apply) {
         try {
             List<org.spongepowered.api.item.enchantment.Enchantment> appliedEnchantments = this.container.get(Keys.APPLIED_ENCHANTMENTS)
+                    .map(LinkedList::new)
                     .orElse(new LinkedList<>());
             List<org.spongepowered.api.item.enchantment.Enchantment> appliedChanges = apply.apply(appliedEnchantments);
             this.setList(Keys.APPLIED_ENCHANTMENTS, appliedChanges);
 
             if (this.container.supports(Keys.STORED_ENCHANTMENTS)) {
-                var storedEnchantments = this.container.get(Keys.STORED_ENCHANTMENTS).orElse(new LinkedList<>());
+                var storedEnchantments = this.container.get(Keys.STORED_ENCHANTMENTS).map(LinkedList::new).orElse(new LinkedList<>());
                 var storedChanges = apply.apply(storedEnchantments);
                 this.setList(Keys.STORED_ENCHANTMENTS, storedChanges);
             }
@@ -459,10 +475,9 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
 
     @Override
     public boolean addAttributeModifier(@NotNull Attribute attribute, @NotNull AttributeModifier modifier) {
-        throw NotImplementedException.createByLazy(ItemMeta.class,
-                "addAttributeModifiers",
-                Attribute.class,
-                AttributeModifier.class);
+        var type = SoakAttributeMap.toSponge(attribute);
+        var spongeModifier = SoakAttributeMap.toSponge(modifier);
+        return this.addAttribute(type, spongeModifier, SoakEquipmentMap.toSponge(modifier.getSlotGroup()));
     }
 
     @Override
@@ -495,23 +510,35 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
     }
 
     @Override
+    @Deprecated
     public Set<Material> getCanDestroy() {
-        return this.container.get(Keys.BREAKABLE_BLOCK_TYPES).orElse(Set.of()).stream().map(Material::getBlockMaterial).collect(Collectors.toSet());
+        return this.container.get(Keys.BREAKABLE_BLOCK_TYPES).orElse(Set.of()).stream().map(SoakBlockMap::toBukkit).collect(Collectors.toSet());
     }
 
     @Override
+    @Deprecated
     public void setCanDestroy(Set<Material> canDestroy) {
-        setSet(Keys.BREAKABLE_BLOCK_TYPES, canDestroy.stream().map(mat -> mat.asBlock()).filter(op -> op.isPresent()).map(op -> op.get()).collect(Collectors.toSet()));
+        setSet(Keys.BREAKABLE_BLOCK_TYPES, canDestroy
+                .stream()
+                .map(SoakBlockMap::toSponge)
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .collect(Collectors.toSet()));
     }
 
     @Override
     public Set<Material> getCanPlaceOn() {
-        return this.container.get(Keys.PLACEABLE_BLOCK_TYPES).orElse(Set.of()).stream().map(Material::getBlockMaterial).collect(Collectors.toSet());
+        return this.container.get(Keys.PLACEABLE_BLOCK_TYPES).orElse(Set.of()).stream().map(SoakBlockMap::toBukkit).collect(Collectors.toSet());
     }
 
     @Override
     public void setCanPlaceOn(Set<Material> canPlaceOn) {
-        setSet(Keys.PLACEABLE_BLOCK_TYPES, canPlaceOn.stream().map(mat -> mat.asBlock()).filter(op -> op.isPresent()).map(op -> op.get()).collect(Collectors.toSet()));
+        setSet(Keys.PLACEABLE_BLOCK_TYPES, canPlaceOn
+                .stream()
+                .map(SoakBlockMap::toSponge)
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .collect(Collectors.toSet()));
     }
 
     @Override
@@ -547,13 +574,7 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
 
     @Override
     public @NotNull Map<String, Object> serialize() {
-        var container = this.asSnapshot().orElseGet(() -> this.asStack().map(ItemStack::createSnapshot).orElseThrow(() -> new IllegalStateException("Unknown mapping"))).toContainer();
-        return container
-                .values(true)
-                .entrySet()
-                .stream()
-                .map(entry -> Map.entry(entry.getKey().asString('.'), entry.getValue()))
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+        return ItemMetaSerializer.serialize(this);
     }
 
     @Override
@@ -590,5 +611,249 @@ public abstract class AbstractItemMeta implements ItemMeta, Damageable {
     @Override
     public @NotNull String getAsString() {
         throw NotImplementedException.createByLazy(ItemMeta.class, "getAsString");
+    }
+
+    @Override
+    public boolean hasDamageValue() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "hasDamageValue");
+    }
+
+    @Override
+    public void resetDamage() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "resetDamage");
+    }
+
+    @Override
+    public boolean hasMaxDamage() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "hasMaxDamage");
+    }
+
+    @Override
+    public int getMaxDamage() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "getMaxDamage");
+    }
+
+    @Override
+    public void setMaxDamage(@Nullable Integer integer) {
+        set(Keys.MAX_DURABILITY, integer);
+    }
+
+    public void manipulate(Function<ItemStackLike, ItemStackLike> to) {
+        this.container = to.apply(this.container);
+    }
+
+    @Override
+    public boolean hasItemName() {
+        return this.container.get(Keys.ITEM_NAME).isPresent();
+    }
+
+    @Override
+    public @NotNull Component itemName() {
+        return this.container.get(Keys.ITEM_NAME).orElseGet(() -> this.container.type().asComponent());
+    }
+
+    @Override
+    public void itemName(@Nullable Component component) {
+        set(Keys.ITEM_NAME, component);
+    }
+
+    @Override
+    public @NotNull String getItemName() {
+        return SoakMessageMap.mapToBukkit(itemName());
+    }
+
+    @Override
+    public void setItemName(@Nullable String s) {
+        if (s == null) {
+            itemName(null);
+        }
+        itemName(SoakMessageMap.toComponent(s));
+    }
+
+    @Override
+    public void removeEnchantments() {
+        setList(Keys.APPLIED_ENCHANTMENTS, Collections.emptyList());
+        setList(Keys.STORED_ENCHANTMENTS, Collections.emptyList());
+    }
+
+    @Override
+    public boolean isHideTooltip() {
+        return this.container.get(Keys.HIDE_TOOLTIP).orElse(false);
+    }
+
+    @Override
+    public void setHideTooltip(boolean b) {
+        set(Keys.HIDE_TOOLTIP, b);
+    }
+
+    @Override
+    public boolean hasEnchantmentGlintOverride() {
+        return this.sponge().get(Keys.ENCHANTMENT_GLINT_OVERRIDE).isPresent();
+    }
+
+    @Override
+    public @NotNull Boolean getEnchantmentGlintOverride() {
+        return this.sponge().get(Keys.ENCHANTMENT_GLINT_OVERRIDE).orElse(false);
+    }
+
+    @Override
+    public void setEnchantmentGlintOverride(@Nullable Boolean aBoolean) {
+        this.set(Keys.ENCHANTMENT_GLINT_OVERRIDE, aBoolean);
+    }
+
+    @Override
+    public boolean isFireResistant() {
+        return this.container.get(Keys.FIRE_RESISTANT).orElse(false);
+    }
+
+    @Override
+    public void setFireResistant(boolean b) {
+        set(Keys.FIRE_RESISTANT, b);
+    }
+
+    @Override
+    public boolean hasMaxStackSize() {
+        return this.container.get(Keys.MAX_STACK_SIZE).isPresent();
+    }
+
+    @Override
+    public int getMaxStackSize() {
+        return this.container.getInt(Keys.MAX_STACK_SIZE).orElseThrow(() -> new RuntimeException("Item '" + this.container.type().key(RegistryTypes.ITEM_TYPE).asString() + "' does not have a max stack size?"));
+    }
+
+    @Override
+    public void setMaxStackSize(@Nullable Integer integer) {
+        set(Keys.MAX_STACK_SIZE, integer);
+    }
+
+    @Override
+    public boolean hasRarity() {
+        return this.container.get(Keys.ITEM_RARITY).isPresent();
+    }
+
+    @Override
+    public @NotNull ItemRarity getRarity() {
+        return this.container.get(Keys.ITEM_RARITY).map(SoakItemStackMap::toBukkit).orElseThrow();
+    }
+
+    @Override
+    public void setRarity(@Nullable ItemRarity itemRarity) {
+        set(Keys.ITEM_RARITY, itemRarity == null ? null : SoakItemStackMap.toSponge(itemRarity));
+    }
+
+    @Override
+    public boolean hasFood() {
+        return sponge().get(Keys.SATURATION).isPresent();
+    }
+
+    @Override
+    public @NotNull FoodComponent getFood() {
+        return new SoakFoodComponent(this);
+    }
+
+    @Override
+    public void setFood(@Nullable FoodComponent foodComponent) {
+        if (foodComponent == null) {
+            remove(Keys.SATURATION);
+            remove(Keys.FOOD_CONVERTS_TO);
+            remove(Keys.EATING_TIME);
+            remove(Keys.CAN_ALWAYS_EAT);
+            return;
+        }
+        set(Keys.CAN_ALWAYS_EAT, foodComponent.canAlwaysEat());
+        set(Keys.FOOD_CONVERTS_TO, foodComponent.getUsingConvertsTo() == null ? null : SoakItemStackMap.toSponge(foodComponent.getUsingConvertsTo()));
+        set(Keys.EATING_TIME, Ticks.of((long) foodComponent.getEatSeconds() * 20));
+        set(Keys.SATURATION, (double) foodComponent.getSaturation());
+    }
+
+    @Override
+    public boolean hasTool() {
+        return this.container.get(Keys.TOOL_DAMAGE_PER_BLOCK).isPresent();
+    }
+
+    @Override
+    public @NotNull ToolComponent getTool() {
+        return new SoakToolComponent(this);
+    }
+
+    @Override
+    public void setTool(@Nullable ToolComponent toolComponent) {
+        var tool = getTool();
+        if (toolComponent == null) {
+            this.remove(Keys.EFFICIENCY);
+            this.remove(Keys.TOOL_RULES);
+            this.remove(Keys.TOOL_DAMAGE_PER_BLOCK);
+            return;
+        }
+        tool.setRules(toolComponent.getRules());
+        tool.setDamagePerBlock(toolComponent.getDamagePerBlock());
+        tool.setDefaultMiningSpeed(toolComponent.getDefaultMiningSpeed());
+    }
+
+    @Override
+    public boolean hasJukeboxPlayable() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "hasJukeboxPlayable");
+    }
+
+    @Override
+    public @NotNull JukeboxPlayableComponent getJukeboxPlayable() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "getJukeboxPlayable");
+    }
+
+    @Override
+    public void setJukeboxPlayable(@Nullable JukeboxPlayableComponent jukeboxPlayableComponent) {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "setJukeboxPlayable", JukeboxPlayableComponent.class);
+    }
+
+    @Override
+    public @NotNull String getAsComponentString() {
+        throw NotImplementedException.createByLazy(ItemMeta.class, "getAsComponentString");
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (!(obj instanceof AbstractItemMeta compareMeta)) {
+            return false;
+        }
+        var compareStack = compareMeta.asStack();
+        var stack = this.asStack();
+        return compareStack.equalTo(stack);
+    }
+
+    public boolean equalsIgnoreQuantity(@Nullable Object object) {
+        if (!(object instanceof AbstractItemMeta compareMeta)) {
+            return false;
+        }
+        var compareStack = compareMeta.asStack();
+        var stack = this.asStack();
+
+        compareStack.setQuantity(1);
+        stack.setQuantity(1);
+        if(compareStack.equalTo(stack)){
+            return true;
+        }
+        //attempt to find difference -> used for debugging
+
+        var originalContainer = stack.asImmutable().toContainer();
+        var compareContainer = compareStack.asImmutable().toContainer();
+
+        var keys =  compareContainer.keys(true);
+        for(var key : keys){
+            if(!originalContainer.contains(key)){
+                return false;
+            }
+            var originalValue = originalContainer.get(key).orElseThrow();
+            if(originalValue instanceof DataView){
+                //deep compare, another key will get inside this view
+                continue;
+            }
+            var compareValue = compareContainer.get(key).orElseThrow();
+            if(!originalValue.equals(compareValue)){
+                return false;
+            }
+        }
+
+
+        return true;
     }
 }
