@@ -15,7 +15,7 @@ import org.spongepowered.api.data.DataHolder;
 import java.util.Optional;
 import java.util.Set;
 
-public abstract class AbstractPersistentData<DH extends DataHolder> implements PersistentDataContainer {
+public abstract class AbstractPersistentData<DH extends DataHolder> implements PersistentDataContainer, SoakPersistentData {
 
     protected DH holder;
 
@@ -25,6 +25,16 @@ public abstract class AbstractPersistentData<DH extends DataHolder> implements P
 
     public DH getHolder() {
         return this.holder;
+    }
+
+    @Override
+    public BukkitPersistentData wrapper() {
+        return this.holder.get(SoakKeys.BUKKIT_DATA).orElseGet(BukkitPersistentData::new);
+    }
+
+    @Override
+    public void copyTo(PersistentDataContainer persistentDataContainer, boolean replace) {
+        throw NotImplementedException.createByLazy(PersistentDataContainer.class, "copyTo", PersistentDataContainer.class, boolean.class);
     }
 
     @Override
@@ -39,7 +49,7 @@ public abstract class AbstractPersistentData<DH extends DataHolder> implements P
 
     @Override
     public @NotNull PersistentDataAdapterContext getAdapterContext() {
-        throw NotImplementedException.createByLazy(PersistentDataContainer.class, "getAdapterContext");
+        return new SoakPersistentDataContext();
     }
 
     @Override
